refactor: split 4,959-line lib.rs into 12 focused modules (#1173)
