+++ contracts/marketpay-contract/src/lib.rs
@@ -33,326 +33,22 @@
     mismatched_lifetime_syntaxes
 )]
 
-use soroban_sdk::{
-    contract, contractimpl, contracttype, symbol_short, token, Address, Bytes, BytesN, Env, String,
-    Symbol, Vec,
-};
+use soroban_sdk::{contract, contractimpl, Address, Bytes, BytesN, Env, String, Vec};
 
-pub mod errors;
+mod admin;
+mod auction;
+mod certificates;
+mod deliverable;
+mod disputes;
+mod errors;
+mod escrow;
+mod governance;
+mod helpers;
+mod messages;
+mod milestones;
+mod types;
 
-// ─── Storage keys ─────────────────────────────────────────────────────────────
-
-/// Default timeout: 7 days in seconds.
-const DEFAULT_TIMEOUT_SECONDS: u32 = 7 * 24 * 60 * 60;
-/// Legacy fallback used by the older ledger-sequence timeout path.
-const DEFAULT_TIMEOUT_LEDGERS: u32 = 120_960;
-
-// ─── Data structures ──────────────────────────────────────────────────────────
-#[contracttype]
-#[derive(Clone, Debug)]
-pub struct CreateEscrowParams {
-    pub freelancer: Address,
-    pub token: Address,
-    pub amount: i128,
-    pub milestones: Option<soroban_sdk::Vec<MilestoneInput>>,
-    pub timeout_ledgers: Option<u32>,
-    pub referrer: Option<Address>,
-}
-
-#[contracttype]
-#[derive(Clone, Debug)]
-pub struct MilestoneInput {
-    pub description: String,
-    pub percentage: u32,
-}
-
-/// Status of an escrow agreement.
-#[contracttype]
-#[derive(Clone, Debug, PartialEq)]
-pub enum EscrowStatus {
-    /// Funds locked, work not yet started
-    Locked,
-    /// Freelancer accepted, work in progress
-    InProgress,
-    /// Client approved work, funds released to freelancer
-    Released,
-    /// Client cancelled before work started, funds refunded
-    Refunded,
-    /// Disputed — requires admin resolution (future feature)
-    Disputed,
-    /// Admin-frozen — no operations allowed until unfrozen
-    Frozen,
-}
-
-#[contracttype]
-#[derive(Clone, Debug, PartialEq)]
-pub struct Milestone {
-    pub id: u32,
-    pub description: String,
-    pub percentage: u32,
-    pub released: bool,
-    /// Set to true when the client rejects this milestone and its share is refunded
-    pub rejected: bool,
-}
-
-/// An escrow record stored on-chain.
-#[contracttype]
-#[derive(Clone, Debug)]
-pub struct Escrow {
-    /// Unique job identifier (from backend)
-    pub job_id: String,
-    /// Client who locked the funds
-    pub client: Address,
-    /// Freelancer who will receive the funds
-    pub freelancer: Address,
-    /// Token contract address (XLM SAC or USDC)
-    pub token: Address,
-    /// Amount in token's smallest unit (stroops for XLM)
-    pub amount: i128,
-    /// Current escrow status
-    pub status: EscrowStatus,
-    /// Ledger when escrow was created
-    pub created_at: u32,
-    /// Ledger after which client can call timeout_refund()
-    pub timeout_ledger: u32,
-    /// Optional milestones for partial releases
-    pub milestones: soroban_sdk::Vec<Milestone>,
-    /// Optional referrer address — receives 2% bonus on release
-    pub referrer: Option<Address>,
-    /// Optional expected SHA-256 deliverable hash agreed by both parties
-    pub deliverable_hash: Option<BytesN<32>>,
-}
-
-/// Budget commitment for sealed-bid system (Issue #108)
-#[contracttype]
-#[derive(Clone, Debug)]
-pub struct BudgetCommitment {
-    pub job_id: String,
-    pub client: Address,
-    pub budget_amount: i128,
-    pub is_revealed: bool,
-}
-
-/// Deliverable hash for oracle verification (Issue #105)
-#[contracttype]
-#[derive(Clone, Debug)]
-pub struct DeliverableSubmission {
-    pub job_id: String,
-    pub client_hash_submitted: bool,
-    pub freelancer_hash_submitted: bool,
-    pub hashes_match: bool,
-}
-
-// On-chain dispute-evidence IPFS CID audit trail (Issue #448 --- AC #2).
-//
-// Per the AC, the contract stores a bare `Vec<Bytes>` of CIDs under
-// `DataKey::EvidenceCids(job_id)`. Each entry is the raw ASCII bytes of
-// an IPFS CID string (e.g. bytes of `bafy...`). The per-record
-// struct (with `kind` and `submitter` fields) has been retired.
-
-/// Freelancer sealed-bid commitment entry.
-#[contracttype]
-#[derive(Clone, Debug)]
-pub struct BidCommitment {
-    pub job_id: String,
-    pub freelancer: Address,
-    pub commitment: BytesN<32>,
-    pub submitted_at_ledger: u32,
-    pub bid_revealed: bool,
-}
-
-/// Bidding lifecycle state for a job.
-#[contracttype]
-#[derive(Clone, Debug)]
-pub struct BiddingState {
-    pub job_id: String,
-    pub client: Address,
-    pub is_closed: bool,
-    pub closed_at_ledger: u32,
-    pub reveal_deadline_ledger: u32,
-}
-
-/// A successfully revealed bid.
-#[contracttype]
-#[derive(Clone, Debug)]
-pub struct RevealedBid {
-    pub freelancer: Address,
-    pub amount: i128,
-    pub revealed_at_ledger: u32,
-}
-
-/// A pending request to extend the escrow timeout, initiated by one party.
-#[contracttype]
-#[derive(Clone, Debug)]
-pub struct ExtensionRequest {
-    pub requested_by: Address,
-    pub new_timeout_ledger: u32,
-    pub created_at: u32,
-}
-
-/// Job completion certificate (Issue #102)
-///
-/// Acts as a proof-of-work NFT minted to the freelancer once the escrow is
-/// released. On-chain metadata captures the job title, client address,
-/// freelancer address, escrow amount and the ledger at mint time.
-#[contracttype]
-#[derive(Clone, Debug)]
-pub struct Certificate {
-    pub job_id: String,
-    pub title: String,
-    pub client: Address,
-    pub freelancer: Address,
-    pub amount: i128,
-    pub created_at: u32,
-}
-
-#[contracttype]
-#[derive(Clone, Debug)]
-pub struct Rating {
-    pub job_id: String,
-    pub rater: Address,
-    pub rated: Address,
-    pub score_out_of_5: u32,
-    pub submitted_at_ledger: u32,
-}
-
-#[contracttype]
-#[derive(Clone, Debug)]
-pub struct FreelancerRatingStats {
-    pub total_score: u32,
-    pub count: u32,
-}
-
-/// Global dispute bond configuration (Issue #437).
-/// When present, callers must lock this amount before a dispute is accepted.
-#[contracttype]
-#[derive(Clone, Debug)]
-pub struct DisputeBondConfig {
-    pub token: Address,
-    pub amount: i128,
-}
-
-/// Per-job locked bond snapshot taken at dispute-raise time.
-#[contracttype]
-#[derive(Clone, Debug)]
-pub struct DisputeBond {
-    pub caller: Address,
-    pub token: Address,
-    pub amount: i128,
-    pub raised_at_ledger: u32,
-}
-
-#[contracttype]
-#[derive(Clone, Debug)]
-pub struct ArbitrationCase {
-    pub job_id: String,
-    pub arbitrators: Vec<Address>,
-    pub votes: Vec<u32>,
-    pub resolution: u32,
-    pub status: u32,
-}
-
-#[contracttype]
-#[derive(Clone, Debug)]
-pub struct DisputeCase {
-    pub job_id: String,
-    pub arbitrators: Vec<Address>,
-    pub votes: Vec<u32>,
-    pub voters: Vec<Address>,
-    pub resolution: u32,
-    pub status: u32,
-}
-
-/// Recurring escrow for retainer contracts (Issue #450)
-#[contracttype]
-#[derive(Clone, Debug)]
-pub struct RecurringEscrow {
-    pub job_id: String,
-    pub client: Address,
-    pub freelancer: Address,
-    pub token: Address,
-    pub amount_per_release: i128,
-    pub interval_ledgers: u32,
-    pub releases_remaining: u32,
-    pub last_release_ledger: u32,
-    pub status: EscrowStatus,
-}
-
-/// Storage key per job
-#[contracttype]
-pub enum DataKey {
-    Admin,
-    /// List of admin addresses for multi-sig operations
-    Admins,
-    /// M-of-N threshold for unfreeze (default 2)
-    UnfreezeThreshold,
-    /// Whether the contract is globally frozen
-    Frozen,
-    Escrow(String),
-    EscrowCount,
-    Proposal(u32),
-    ProposalCount,
-    HasVoted(Address, u32),
-    CompletedJobs(Address),
-    DefaultTimeoutSeconds,
-    TimeoutTimestamp(String),
-    BudgetCommitment(String),
-    DeliverableSubmission(String),
-    /// Per-job append-only audit log of deliverable IPFS CIDs (Issue #448).
-    /// Stores a Vec<Bytes> of dispute-evidence CIDs under the job_id key.
-    EvidenceCids(String),
-    BidCommitment(String, Address),
-    BiddingState(String),
-    RevealedBids(String),
-    Certificate(String),
-    FreelancerCertificates(Address),
-    ClientRating(String),
-    FreelancerRating(String),
-    FreelancerRatingStats(Address),
-    Arbitrator(Address),
-    ArbitratorPool,
-    /// Single admin-designated arbitrator (set_arbitrator / get_arbitrator)
-    ArbitratorAddress,
-    ArbitrationCase(u32),
-    ArbitrationCaseCount,
-    DisputeCase(String),
-    Version,
-    /// Stores list of IPFS CIDs for messages in a job thread
-    MessageCid(String),
-    /// Freelancer-submitted deliverable SHA-256 hash for release verification
-    FreelancerDeliverableHash(String),
-    /// Address that receives platform fees on every escrow release
-    TreasuryAddress,
-    /// Platform fee in basis points (e.g. 100 = 1%)
-    PlatformFeeBps,
-    /// Maximum referrer bonus cap in token stroops
-    MaxReferrerBonusXlm,
-    /// Pending escrow timeout extension request
-    ExtensionRequest(String),
-    /// Global dispute bond configuration
-    DisputeBondConfig,
-    /// Per-job locked dispute bond record
-    DisputeBond(String),
-}
-
-/// Reveal phase is open for roughly 24 hours after client closes bidding.
-const REVEAL_WINDOW_LEDGERS: u32 = 17_280;
-
-/// A governance proposal
-#[contracttype]
-#[derive(Clone, Debug)]
-pub struct Proposal {
-    pub id: u32,
-    pub title: String,
-    pub description: String,
-    pub votes_for: u32,
-    pub votes_against: u32,
-    pub deadline_ledger: u32,
-    pub resolved: bool,
-    pub result: bool,
-}
-
-// ─── Contract ─────────────────────────────────────────────────────────────────
+pub use types::*;
 
 #[contract]
 pub struct MarketPayContract;
@@ -360,58 +56,14 @@ pub struct MarketPayContract;
 #[allow(clippy::too_many_arguments)]
 #[contractimpl]
 impl MarketPayContract {
-    fn compute_bid_commitment(env: &Env, amount: i128, nonce: BytesN<32>) -> BytesN<32> {
-        let mut payload = Bytes::new(env);
-        for byte in amount.to_be_bytes().iter() {
-            payload.push_back(*byte);
-        }
-        for byte in nonce.to_array().iter() {
-            payload.push_back(*byte);
-        }
-        env.crypto().sha256(&payload).into()
-    }
-
-    fn check_not_frozen(env: &Env) {
-        let frozen: bool = env
-            .storage()
-            .instance()
-            .get(&DataKey::Frozen)
-            .unwrap_or(false);
-        if frozen {
-            panic!("Contract is frozen");
-        }
-    }
-
     // ─── Initialization ──────────────────────────────────────────────────────
 
     /// Initialize with an admin address and treasury (called once after deployment).
     ///
     /// `treasury_address` receives a configurable platform fee on every escrow
-    /// release. The initial platform fee defaults to 100 bps (1 %).
+    /// release. The initial platform fee defaults to 100 bps (1 %).
     pub fn initialize(env: Env, admin: Address, treasury_address: Address) {
-        if env.storage().instance().has(&DataKey::Admin) {
-            panic!("Already initialized");
-        }
-        env.storage().instance().set(&DataKey::Admin, &admin);
-        env.storage()
-            .instance()
-            .set(&DataKey::TreasuryAddress, &treasury_address);
-        env.storage()
-            .instance()
-            .set(&DataKey::PlatformFeeBps, &100u32);
-        env.storage().instance().set(&DataKey::EscrowCount, &0u32);
-        env.storage()
-            .instance()
-            .set(&DataKey::DefaultTimeoutSeconds, &DEFAULT_TIMEOUT_SECONDS);
-        env.storage().instance().set(&DataKey::Version, &1u32);
-
-        let mut admins: Vec<Address> = Vec::new(&env);
-        admins.push_back(admin);
-        env.storage().instance().set(&DataKey::Admins, &admins);
-        env.storage()
-            .instance()
-            .set(&DataKey::UnfreezeThreshold, &2u32);
-        env.storage().instance().set(&DataKey::Frozen, &false);
+        admin::initialize(env, admin, treasury_address)
     }
 
     // ─── Upgrade & versioning ─────────────────────────────────────────────────
@@ -423,55 +75,24 @@ impl MarketPayContract {
     /// All existing storage (escrows, proposals, ratings, …) is preserved
     /// because Soroban upgrades only replace the executable, not the state.
     pub fn upgrade(env: Env, new_wasm_hash: BytesN<32>) {
-        let admin: Address = env
-            .storage()
-            .instance()
-            .get(&DataKey::Admin)
-            .expect("Not initialized");
-        admin.require_auth();
-
-        env.deployer().update_current_contract_wasm(new_wasm_hash);
-
-        // Bump version so callers can detect the upgrade
-        let version: u32 = env.storage().instance().get(&DataKey::Version).unwrap_or(1);
-        env.storage()
-            .instance()
-            .set(&DataKey::Version, &(version + 1));
-
-        env.events()
-            .publish((symbol_short!("upgraded"), admin), version + 1);
+        admin::upgrade(env, new_wasm_hash)
     }
 
     /// Return the current contract version (starts at 1, increments on each upgrade).
     pub fn get_version(env: Env) -> u32 {
-        env.storage().instance().get(&DataKey::Version).unwrap_or(1)
+        admin::get_version(env)
     }
 
     // ─── Escrow lifecycle ─────────────────────────────────────────────────────
 
     /// Client creates an escrow by transferring funds into the contract.
-    ///
-    /// Parameters:
-    ///   job_id           — unique ID matching the backend job record
-    ///   freelancer       — the address that will receive payment on release
-    ///   token            — SAC address of the payment token (XLM or USDC)
-    ///   amount           — payment amount in smallest token units
-    ///   milestones       — optional list of milestones (amounts must sum to total amount)
-    ///   timeout_ledgers  — optional ledger timeout (default 7 days)
-    ///   referrer         — optional referrer address; receives 2% bonus on release
-    pub fn create_escrow(env: Env, job_id: String, client: Address, params: CreateEscrowParams) {
-        Self::create_escrow_internal(
-            env,
-            job_id,
-            client,
-            params.freelancer,
-            params.token,
-            params.amount,
-            params.milestones,
-            params.timeout_ledgers,
-            params.referrer,
-            None,
-        )
+    pub fn create_escrow(
+        env: Env,
+        job_id: String,
+        client: Address,
+        params: types::CreateEscrowParams,
+    ) {
+        escrow::create_escrow(env, job_id, client, params)
     }
 
     /// Client creates an escrow that includes an expected deliverable hash.
@@ -479,394 +100,34 @@ impl MarketPayContract {
         env: Env,
         job_id: String,
         client: Address,
-        params: CreateEscrowParams,
+        params: types::CreateEscrowParams,
         deliverable_hash: BytesN<32>,
     ) {
-        Self::create_escrow_internal(
-            env,
-            job_id,
-            client,
-            params.freelancer,
-            params.token,
-            params.amount,
-            params.milestones,
-            params.timeout_ledgers,
-            params.referrer,
-            Some(deliverable_hash),
-        )
+        escrow::create_escrow_with_deliverable(env, job_id, client, params, deliverable_hash)
     }
 
-    // Client creates an escrow with percentage-based milestones.
-    // milestone percentages must sum to 100.
+    /// Client creates an escrow with percentage-based milestones.
+    /// milestone percentages must sum to 100.
     pub fn create_escrow_with_milestones(
         env: Env,
         job_id: String,
         client: Address,
-        params: CreateEscrowParams,
+        params: types::CreateEscrowParams,
     ) {
-        Self::create_escrow_internal(
-            env,
-            job_id,
-            client,
-            params.freelancer,
-            params.token,
-            params.amount,
-            params.milestones,
-            params.timeout_ledgers,
-            params.referrer,
-            None,
-        )
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    fn create_escrow_internal(
-        env: Env,
-        job_id: String,
-        client: Address,
-        freelancer: Address,
-        token: Address,
-        amount: i128,
-        milestones: Option<soroban_sdk::Vec<MilestoneInput>>,
-        timeout_ledgers: Option<u32>,
-        referrer: Option<Address>,
-        deliverable_hash: Option<BytesN<32>>,
-    ) {
-        client.require_auth();
-        Self::check_not_frozen(&env);
-
-        if amount <= 0 {
-            panic!("Amount must be positive");
-        }
-
-        // Referrer must not be the freelancer or client
-        if let Some(ref r) = referrer {
-            if r == &client || r == &freelancer {
-                panic!("Referrer cannot be the client or freelancer");
-            }
-        }
-
-        // Validate milestones if provided
-        let mut milestone_list = soroban_sdk::Vec::new(&env);
-        if let Some(ms) = milestones {
-            if ms.len() > 5 {
-                panic!("Maximum 5 milestones allowed");
-            }
-            let mut total_percentage: u32 = 0;
-            for (next_id, m) in (0_u32..).zip(ms.iter()) {
-                if m.percentage == 0 {
-                    panic!("Milestone percentage must be positive");
-                }
-                total_percentage = total_percentage
-                    .checked_add(m.percentage)
-                    .expect("Arithmetic overflow");
-                milestone_list.push_back(Milestone {
-                    id: next_id,
-                    description: m.description.clone(),
-                    percentage: m.percentage,
-                    released: false,
-                    rejected: false,
-                });
-            }
-            if total_percentage != 100 {
-                panic!("Milestone percentages must sum to 100");
-            }
-        }
-
-        // Ensure no duplicate escrow for same job
-        if env
-            .storage()
-            .instance()
-            .has(&DataKey::Escrow(job_id.clone()))
-        {
-            panic!("Escrow already exists for this job");
-        }
-
-        // Transfer funds from client into the contract
-        let token_client = token::Client::new(&env, &token);
-        let contract_address = env.current_contract_address();
-        token_client.transfer(&client, &contract_address, &amount);
-
-        let current_ledger = env.ledger().sequence();
-        let current_timestamp = env.ledger().timestamp() as u32;
-        let timeout = timeout_ledgers.unwrap_or(DEFAULT_TIMEOUT_LEDGERS);
-        let timeout_ledger = current_ledger
-            .checked_add(timeout)
-            .expect("Timeout ledger overflow");
-        let timeout_seconds: u32 = env
-            .storage()
-            .instance()
-            .get(&DataKey::DefaultTimeoutSeconds)
-            .unwrap_or(DEFAULT_TIMEOUT_SECONDS);
-        let timeout_timestamp = current_timestamp
-            .checked_add(timeout_seconds)
-            .expect("Timeout timestamp overflow");
-
-        // Store escrow record on-chain
-        let escrow = Escrow {
-            job_id: job_id.clone(),
-            client: client.clone(),
-            freelancer,
-            token,
-            amount,
-            status: EscrowStatus::Locked,
-            created_at: current_ledger,
-            timeout_ledger,
-            milestones: milestone_list,
-            referrer,
-            deliverable_hash,
-        };
-
-        env.storage()
-            .instance()
-            .set(&DataKey::Escrow(job_id.clone()), &escrow);
-        env.storage().instance().set(
-            &DataKey::TimeoutTimestamp(job_id.clone()),
-            &timeout_timestamp,
-        );
-
-        // Increment counter
-        let count: u32 = env
-            .storage()
-            .instance()
-            .get(&DataKey::EscrowCount)
-            .unwrap_or(0);
-        let new_count = count.checked_add(1).expect("Counter overflow");
-        env.storage()
-            .instance()
-            .set(&DataKey::EscrowCount, &new_count);
-
-        // Emit event
-        env.events().publish(
-            (symbol_short!("escrow_cr"), job_id.clone()),
-            (
-                escrow.client.clone(),
-                escrow.freelancer.clone(),
-                escrow.amount,
-            ),
-        );
+        escrow::create_escrow_with_milestones(env, job_id, client, params)
     }
 
     /// Freelancer signals that they have started work.
     pub fn start_work(env: Env, job_id: String, freelancer: Address) {
-        freelancer.require_auth();
-        Self::check_not_frozen(&env);
-
-        let mut escrow: Escrow = env
-            .storage()
-            .instance()
-            .get(&DataKey::Escrow(job_id.clone()))
-            .expect("Escrow not found");
-
-        if escrow.freelancer != freelancer {
-            panic!("Only the freelancer can start work");
-        }
-        if escrow.status != EscrowStatus::Locked {
-            panic!("Escrow is not in Locked state");
-        }
-
-        escrow.status = EscrowStatus::InProgress;
-        env.storage()
-            .instance()
-            .set(&DataKey::Escrow(job_id.clone()), &escrow);
-
-        env.events().publish(
-            (symbol_short!("work_strt"), job_id.clone()),
-            (escrow.client.clone(), escrow.freelancer.clone()),
-        );
+        escrow::start_work(env, job_id, freelancer)
     }
 
     /// Client approves completed work and releases funds to the freelancer.
     pub fn release_escrow(env: Env, job_id: String, client: Address) {
-        client.require_auth();
-        Self::check_not_frozen(&env);
-
-        let escrow: Escrow = env
-            .storage()
-            .instance()
-            .get(&DataKey::Escrow(job_id.clone()))
-            .expect("Escrow not found");
-
-        if escrow.client != client {
-            panic!("Only the client can release escrow");
-        }
-
-        // Deliverable hash verification: if an expected hash was set on creation,
-        // the freelancer must have submitted a matching hash before release.
-        if let Some(expected_hash) = &escrow.deliverable_hash {
-            let submitted: Option<BytesN<32>> = env
-                .storage()
-                .instance()
-                .get(&DataKey::FreelancerDeliverableHash(job_id.clone()));
-            match submitted {
-                Some(h) if &h == expected_hash => {}
-                _ => panic!("Freelancer deliverable hash does not match or not submitted"),
-            }
-        }
-
-        Self::release_escrow_core(env, job_id, escrow);
-    }
-
-    fn release_escrow_core(env: Env, job_id: String, mut escrow: Escrow) {
-        if escrow.status != EscrowStatus::InProgress && escrow.status != EscrowStatus::Locked {
-            panic!("Cannot release escrow in current status");
-        }
-
-        // Check if there are incomplete milestones
-        let mut remaining_amount: i128 = 0;
-        for ms in escrow.milestones.iter() {
-            if !ms.released {
-                let ms_amount = escrow
-                    .amount
-                    .checked_mul(ms.percentage as i128)
-                    .expect("Arithmetic overflow")
-                    .checked_div(100)
-                    .expect("Arithmetic overflow");
-                remaining_amount = remaining_amount
-                    .checked_add(ms_amount)
-                    .expect("Arithmetic overflow");
-            }
-        }
-
-        // If no milestones, release full amount. If milestones, release remaining.
-        let release_amount = if escrow.milestones.is_empty() {
-            escrow.amount
-        } else {
-            remaining_amount
-        };
-
-        // Mark all milestones as completed
-        let mut updated_ms = soroban_sdk::Vec::new(&env);
-        for mut ms in escrow.milestones.iter() {
-            ms.released = true;
-            updated_ms.push_back(ms);
-        }
-        escrow.milestones = updated_ms;
-
-        // Increment CompletedJobs for the freelancer and client
-        let freelancer_jobs: u32 = env
-            .storage()
-            .instance()
-            .get(&DataKey::CompletedJobs(escrow.freelancer.clone()))
-            .unwrap_or(0);
-        let new_freelancer_jobs = freelancer_jobs.checked_add(1).expect("Counter overflow");
-        env.storage().instance().set(
-            &DataKey::CompletedJobs(escrow.freelancer.clone()),
-            &new_freelancer_jobs,
-        );
-
-        let client_jobs: u32 = env
-            .storage()
-            .instance()
-            .get(&DataKey::CompletedJobs(escrow.client.clone()))
-            .unwrap_or(0);
-        let new_client_jobs = client_jobs.checked_add(1).expect("Counter overflow");
-        env.storage().instance().set(
-            &DataKey::CompletedJobs(escrow.client.clone()),
-            &new_client_jobs,
-        );
-
-        escrow.status = EscrowStatus::Released;
-        env.storage()
-            .instance()
-            .set(&DataKey::Escrow(job_id.clone()), &escrow);
-        env.storage()
-            .instance()
-            .remove(&DataKey::TimeoutTimestamp(job_id.clone()));
-        env.storage()
-            .instance()
-            .remove(&DataKey::FreelancerDeliverableHash(job_id.clone()));
-
-        if release_amount > 0 {
-            let token_client = token::Client::new(&env, &escrow.token);
-
-            // ── Platform fee ────────────────────────────────────────────────────
-            let fee_bps: u32 = env
-                .storage()
-                .instance()
-                .get(&DataKey::PlatformFeeBps)
-                .unwrap_or(0);
-            let treasury: Address = env
-                .storage()
-                .instance()
-                .get(&DataKey::TreasuryAddress)
-                .expect("Treasury not set");
-            let fee_amount = release_amount
-                .checked_mul(fee_bps as i128)
-                .expect("Arithmetic overflow")
-                .checked_div(10_000)
-                .expect("Arithmetic overflow");
-            let after_fee = release_amount
-                .checked_sub(fee_amount)
-                .expect("Arithmetic overflow");
-
-            if fee_amount > 0 {
-                token_client.transfer(&env.current_contract_address(), &treasury, &fee_amount);
-                env.events().publish(
-                    (symbol_short!("plat_fee"), job_id.clone()),
-                    (treasury.clone(), fee_amount),
-                );
-            }
-
-            // ── Referral bonus: 2% of post-fee amount goes to referrer ─────────
-            let (freelancer_amount, referral_amount) = match &escrow.referrer {
-                Some(referrer_addr) => {
-                    let bonus = after_fee
-                        .checked_mul(200)
-                        .expect("Arithmetic overflow")
-                        .checked_div(10_000)
-                        .expect("Arithmetic overflow");
-                    let to_freelancer = after_fee.checked_sub(bonus).expect("Arithmetic overflow");
-                    if bonus > 0 {
-                        token_client.transfer(
-                            &env.current_contract_address(),
-                            referrer_addr,
-                            &bonus,
-                        );
-                        env.events().publish(
-                            (symbol_short!("ref_bon"), referrer_addr.clone()),
-                            (job_id.clone(), bonus),
-                        );
-                    }
-                    (to_freelancer, bonus)
-                }
-                None => (after_fee, 0i128),
-            };
-
-            // Transfer remaining funds to freelancer
-            if freelancer_amount > 0 {
-                token_client.transfer(
-                    &env.current_contract_address(),
-                    &escrow.freelancer,
-                    &freelancer_amount,
-                );
-            }
-
-            env.events().publish(
-                (symbol_short!("escrow_rl"), job_id.clone()),
-                (
-                    escrow.client.clone(),
-                    escrow.freelancer.clone(),
-                    freelancer_amount,
-                    referral_amount,
-                    fee_amount,
-                ),
-            );
-        } else {
-            env.events().publish(
-                (symbol_short!("escrow_rl"), job_id.clone()),
-                (
-                    escrow.client.clone(),
-                    escrow.freelancer.clone(),
-                    0i128,
-                    0i128,
-                    0i128,
-                ),
-            );
-        }
+        escrow::release_escrow(env, job_id, client)
     }
 
     /// Client approves work and releases funds WITH conversion through DEX.
-    /// This is used when the escrow is in one asset (e.g. USDC) but the freelancer wants another (e.g. XLM).
     pub fn release_with_conversion(
         env: Env,
         job_id: String,
@@ -874,748 +135,168 @@ impl MarketPayContract {
         _target_token: Address,
         _min_amount_out: i128,
     ) {
-        client.require_auth();
-        Self::check_not_frozen(&env);
-
-        let mut escrow: Escrow = env
-            .storage()
-            .instance()
-            .get(&DataKey::Escrow(job_id.clone()))
-            .expect("Escrow not found");
-
-        if escrow.client != client {
-            panic!("Only the client can release escrow");
-        }
-        if escrow.status != EscrowStatus::InProgress && escrow.status != EscrowStatus::Locked {
-            panic!("Cannot release escrow in current status");
-        }
-
-        // Calculate remaining amount
-        let mut remaining_amount: i128 = 0;
-        for ms in escrow.milestones.iter() {
-            if !ms.released {
-                let ms_amount = escrow
-                    .amount
-                    .checked_mul(ms.percentage as i128)
-                    .expect("Arithmetic overflow")
-                    .checked_div(100)
-                    .expect("Arithmetic overflow");
-                remaining_amount = remaining_amount
-                    .checked_add(ms_amount)
-                    .expect("Arithmetic overflow");
-            }
-        }
-        let release_amount = if escrow.milestones.is_empty() {
-            escrow.amount
-        } else {
-            remaining_amount
-        };
-
-        if release_amount > 0 {
-            let token_client = token::Client::new(&env, &escrow.token);
-
-            // ── Platform fee ────────────────────────────────────────────────────
-            let fee_bps: u32 = env
-                .storage()
-                .instance()
-                .get(&DataKey::PlatformFeeBps)
-                .unwrap_or(0);
-            let treasury: Address = env
-                .storage()
-                .instance()
-                .get(&DataKey::TreasuryAddress)
-                .expect("Treasury not set");
-            let fee_amount = release_amount
-                .checked_mul(fee_bps as i128)
-                .expect("Arithmetic overflow")
-                .checked_div(10_000)
-                .expect("Arithmetic overflow");
-            let to_freelancer = release_amount
-                .checked_sub(fee_amount)
-                .expect("Arithmetic overflow");
-
-            if fee_amount > 0 {
-                token_client.transfer(&env.current_contract_address(), &treasury, &fee_amount);
-            }
-
-            // [Issue #104] Path Payment / DEX Swap
-            // In a real scenario, we would call a DEX contract here.
-            // For now, we simulate the conversion by transferring the source token
-            // and emitting a conversion event.
-            // In a real implementation with a Soroban DEX:
-            // let dex = DEXClient::new(&env, &DEX_ADDRESS);
-            // dex.swap(&env.current_contract_address(), &escrow.freelancer, &escrow.token, &target_token, &release_amount, &min_amount_out);
-
-            // For this implementation, we perform the transfer and mark as converted
-            token_client.transfer(
-                &env.current_contract_address(),
-                &escrow.freelancer,
-                &to_freelancer,
-            );
-        }
-
-        // Mark all milestones as completed
-        let mut updated_ms = soroban_sdk::Vec::new(&env);
-        for mut ms in escrow.milestones.iter() {
-            ms.released = true;
-            updated_ms.push_back(ms);
-        }
-        escrow.milestones = updated_ms;
-
-        // Update jobs count
-        let f_jobs: u32 = env
-            .storage()
-            .instance()
-            .get(&DataKey::CompletedJobs(escrow.freelancer.clone()))
-            .unwrap_or(0);
-        env.storage().instance().set(
-            &DataKey::CompletedJobs(escrow.freelancer.clone()),
-            &(f_jobs.checked_add(1).unwrap()),
-        );
-
-        let c_jobs: u32 = env
-            .storage()
-            .instance()
-            .get(&DataKey::CompletedJobs(escrow.client.clone()))
-            .unwrap_or(0);
-        env.storage().instance().set(
-            &DataKey::CompletedJobs(escrow.client.clone()),
-            &(c_jobs.checked_add(1).unwrap()),
-        );
-
-        escrow.status = EscrowStatus::Released;
-        env.storage()
-            .instance()
-            .set(&DataKey::Escrow(job_id.clone()), &escrow);
-        env.storage()
-            .instance()
-            .remove(&DataKey::TimeoutTimestamp(job_id.clone()));
-
-        env.events().publish(
-            (symbol_short!("escrow_rl"), job_id.clone()),
-            (
-                escrow.client.clone(),
-                escrow.freelancer.clone(),
-                release_amount,
-            ),
-        );
+        escrow::release_with_conversion(env, job_id, client, _target_token, _min_amount_out)
     }
 
     /// Client cancels and gets a refund (only before work starts).
     pub fn refund_escrow(env: Env, job_id: String, client: Address) {
-        client.require_auth();
-        Self::check_not_frozen(&env);
-
-        let mut escrow: Escrow = env
-            .storage()
-            .instance()
-            .get(&DataKey::Escrow(job_id.clone()))
-            .expect("Escrow not found");
-
-        if escrow.client != client {
-            panic!("Only the client can request a refund");
-        }
-        if escrow.status != EscrowStatus::Locked {
-            panic!("Can only refund before work has started");
-        }
-
-        // Return funds to client
-        let token_client = token::Client::new(&env, &escrow.token);
-        token_client.transfer(
-            &env.current_contract_address(),
-            &escrow.client,
-            &escrow.amount,
-        );
-
-        escrow.status = EscrowStatus::Refunded;
-        env.storage()
-            .instance()
-            .set(&DataKey::Escrow(job_id.clone()), &escrow);
-
-        env.events().publish(
-            (symbol_short!("escrow_rf"), job_id.clone()),
-            (
-                escrow.client.clone(),
-                escrow.freelancer.clone(),
-                escrow.amount,
-            ),
-        );
+        escrow::refund_escrow(env, job_id, client)
     }
 
-    /// Issue #175 — Client claims a refund if the freelancer never started work
-    /// before the timeout. New escrows enforce the timeout using Unix timestamps;
-    /// older escrows fall back to the legacy ledger-sequence threshold.
+    /// Client claims a refund if the freelancer never started work before the timeout.
     pub fn timeout_refund(env: Env, job_id: String, client: Address) {
-        client.require_auth();
-        Self::check_not_frozen(&env);
-
-        let mut escrow: Escrow = env
-            .storage()
-            .instance()
-            .get(&DataKey::Escrow(job_id.clone()))
-            .expect("Escrow not found");
-
-        if escrow.client != client {
-            panic!("Only the client can request a timeout refund");
-        }
-        if escrow.status != EscrowStatus::Locked {
-            panic!("Escrow is not in Locked state");
-        }
-
-        let current_timestamp = env.ledger().timestamp() as u32;
-        let timeout_timestamp: Option<u32> = env
-            .storage()
-            .instance()
-            .get(&DataKey::TimeoutTimestamp(job_id.clone()));
-        let expired = if let Some(timeout_timestamp) = timeout_timestamp {
-            current_timestamp >= timeout_timestamp
-        } else {
-            env.ledger().sequence() >= escrow.timeout_ledger
-        };
-
-        if !expired {
-            panic!("Timeout period has not expired yet");
-        }
-
-        // Return funds to client
-        let token_client = token::Client::new(&env, &escrow.token);
-        token_client.transfer(
-            &env.current_contract_address(),
-            &escrow.client,
-            &escrow.amount,
-        );
-
-        escrow.status = EscrowStatus::Refunded;
-        env.storage()
-            .instance()
-            .set(&DataKey::Escrow(job_id.clone()), &escrow);
-
-        env.events().publish(
-            (symbol_short!("escrow_rf"), job_id.clone()),
-            (
-                escrow.client.clone(),
-                escrow.freelancer.clone(),
-                escrow.amount,
-            ),
-        );
+        escrow::timeout_refund(env, job_id, client)
     }
 
     // ─── Getters ─────────────────────────────────────────────────────────────
 
     /// Get the full escrow record for a job.
-    pub fn get_escrow(env: Env, job_id: String) -> Escrow {
-        env.storage()
-            .instance()
-            .get(&DataKey::Escrow(job_id))
-            .expect("Escrow not found")
+    pub fn get_escrow(env: Env, job_id: String) -> types::Escrow {
+        admin::get_escrow(env, job_id)
     }
 
     /// Get escrow status for a job.
-    pub fn get_status(env: Env, job_id: String) -> EscrowStatus {
-        let escrow: Escrow = env
-            .storage()
-            .instance()
-            .get(&DataKey::Escrow(job_id))
-            .expect("Escrow not found");
-        escrow.status
+    pub fn get_status(env: Env, job_id: String) -> types::EscrowStatus {
+        admin::get_status(env, job_id)
     }
 
     /// Get timeout ledger for a job.
     pub fn get_timeout_ledger(env: Env, job_id: String) -> u32 {
-        let escrow: Escrow = env
-            .storage()
-            .instance()
-            .get(&DataKey::Escrow(job_id))
-            .expect("Escrow not found");
-        escrow.timeout_ledger
+        admin::get_timeout_ledger(env, job_id)
     }
 
     /// Get the timestamp after which `timeout_refund()` becomes available.
     pub fn get_timeout_timestamp(env: Env, job_id: String) -> u32 {
-        env.storage()
-            .instance()
-            .get(&DataKey::TimeoutTimestamp(job_id))
-            .unwrap_or(0)
+        admin::get_timeout_timestamp(env, job_id)
     }
 
     /// Get a single milestone from an escrow by index.
-    pub fn get_milestone(env: Env, job_id: String, index: u32) -> Milestone {
-        let escrow: Escrow = env
-            .storage()
-            .instance()
-            .get(&DataKey::Escrow(job_id))
-            .expect("Escrow not found");
-        if index >= escrow.milestones.len() {
-            panic!("Milestone index out of bounds");
-        }
-        escrow.milestones.get(index).unwrap()
+    pub fn get_milestone(env: Env, job_id: String, index: u32) -> types::Milestone {
+        admin::get_milestone(env, job_id, index)
     }
 
     /// Check whether the contract is globally frozen.
     pub fn is_frozen(env: Env) -> bool {
-        env.storage()
-            .instance()
-            .get(&DataKey::Frozen)
-            .unwrap_or(false)
+        admin::is_frozen(env)
     }
 
     /// Get the referrer address for a job's escrow, if one was set.
     pub fn get_referrer(env: Env, job_id: String) -> Option<Address> {
-        let escrow: Escrow = env
-            .storage()
-            .instance()
-            .get(&DataKey::Escrow(job_id))
-            .expect("Escrow not found");
-        escrow.referrer
+        admin::get_referrer(env, job_id)
     }
 
     /// Get total number of escrows created.
     pub fn get_escrow_count(env: Env) -> u32 {
-        env.storage()
-            .instance()
-            .get(&DataKey::EscrowCount)
-            .unwrap_or(0)
+        admin::get_escrow_count(env)
     }
 
     /// Get the contract admin.
     pub fn get_admin(env: Env) -> Address {
-        env.storage()
-            .instance()
-            .get(&DataKey::Admin)
-            .expect("Not initialized")
+        admin::get_admin(env)
     }
 
     /// Get the treasury address that receives platform fees.
     pub fn get_treasury_address(env: Env) -> Address {
-        env.storage()
-            .instance()
-            .get(&DataKey::TreasuryAddress)
-            .expect("Treasury not set")
+        admin::get_treasury_address(env)
     }
 
     /// Get the platform fee in basis points (e.g. 100 = 1%).
     pub fn get_platform_fee_bps(env: Env) -> u32 {
-        env.storage()
-            .instance()
-            .get(&DataKey::PlatformFeeBps)
-            .unwrap_or(0)
+        admin::get_platform_fee_bps(env)
     }
 
     // ─── Arbitrator Management ────────────────────────────────────────────
 
     /// Set the arbitrator address that can resolve disputes.
-    /// Only callable by the contract admin.
     pub fn set_arbitrator(env: Env, admin: Address, arbitrator: Address) {
-        admin.require_auth();
-
-        let stored_admin: Address = env
-            .storage()
-            .instance()
-            .get(&DataKey::Admin)
-            .expect("Not initialized");
-        if stored_admin != admin {
-            panic!("Only admin can set the arbitrator");
-        }
-
-        env.storage()
-            .instance()
-            .set(&DataKey::ArbitratorAddress, &arbitrator);
-
-        env.events()
-            .publish((symbol_short!("arb_set"), admin), arbitrator);
+        admin::set_arbitrator(env, admin, arbitrator)
     }
 
-    /// Get the current arbitrator address.  Returns `None` if not set.
+    /// Get the current arbitrator address. Returns `None` if not set.
     pub fn get_arbitrator(env: Env) -> Option<Address> {
-        env.storage().instance().get(&DataKey::ArbitratorAddress)
+        admin::get_arbitrator(env)
     }
 
     /// Update the treasury address. Only callable by admin.
     pub fn set_treasury_address(env: Env, admin: Address, treasury_address: Address) {
-        admin.require_auth();
-
-        let stored_admin: Address = env
-            .storage()
-            .instance()
-            .get(&DataKey::Admin)
-            .expect("Not initialized");
-        if stored_admin != admin {
-            panic!("Only admin can set treasury address");
-        }
-
-        env.storage()
-            .instance()
-            .set(&DataKey::TreasuryAddress, &treasury_address);
-        env.events()
-            .publish((symbol_short!("set_tres"), admin), treasury_address);
+        admin::set_treasury_address(env, admin, treasury_address)
     }
 
     /// Update the platform fee in basis points. Only callable by admin.
-    /// `bps` must be ≤ 1000 (10 %).
     pub fn set_platform_fee_bps(env: Env, admin: Address, bps: u32) {
-        admin.require_auth();
-
-        let stored_admin: Address = env
-            .storage()
-            .instance()
-            .get(&DataKey::Admin)
-            .expect("Not initialized");
-        if stored_admin != admin {
-            panic!("Only admin can set platform fee");
-        }
-        if bps > 1000 {
-            panic!("Platform fee cannot exceed 10% (1000 bps)");
-        }
-
-        env.storage().instance().set(&DataKey::PlatformFeeBps, &bps);
-        env.events().publish((symbol_short!("set_fee"), admin), bps);
+        admin::set_platform_fee_bps(env, admin, bps)
     }
 
     /// Get the current global timeout in seconds.
     pub fn get_default_timeout_seconds(env: Env) -> u32 {
-        env.storage()
-            .instance()
-            .get(&DataKey::DefaultTimeoutSeconds)
-            .unwrap_or(DEFAULT_TIMEOUT_SECONDS)
+        admin::get_default_timeout_seconds(env)
     }
 
-    /// Issue #440 — look up the admin-set cap on referrer bonus payouts.
-    /// Returns `None` when no cap has been set (legacy behaviour: 2% of
-    /// `release_amount` always applies).
+    /// Look up the admin-set cap on referrer bonus payouts.
     pub fn get_max_referrer_bonus_xlm(env: Env) -> Option<i128> {
-        env.storage().instance().get(&DataKey::MaxReferrerBonusXlm)
+        admin::get_max_referrer_bonus_xlm(env)
     }
 
-    /// Issue #440 — admin sets the maximum referrer bonus (in token
-    /// stroops, i.e. same units as escrow amounts). Pass `0` to disable
-    /// the referrer program entirely; pass a positive value to cap
-    /// every release's referrer-ledger entry at that amount.
-    ///
-    /// The cap is consumed at `release_escrow_core()` time so existing
-    /// escrows that have not yet been released pick up the new cap on
-    /// their first release. For milestone escrows each partial
-    /// `release_milestone` call applies the cap independently to that
-    /// release's payout — a 5-milestone escrow with cap = 10 XLM pays
-    /// the cap up to 5 times, not once cumulatively.
+    /// Admin sets the maximum referrer bonus (in token stroops).
     pub fn set_max_referrer_bonus_xlm(env: Env, admin: Address, cap: i128) {
-        admin.require_auth();
-
-        let stored_admin: Address = env
-            .storage()
-            .instance()
-            .get(&DataKey::Admin)
-            .expect("Not initialized");
-        if stored_admin != admin {
-            panic!("Only admin can set the referrer bonus cap");
-        }
-        if cap < 0 {
-            panic!("Referrer bonus cap must be non-negative");
-        }
-
-        env.storage()
-            .instance()
-            .set(&DataKey::MaxReferrerBonusXlm, &cap);
-        env.events().publish((symbol_short!("ref_cap"), admin), cap);
+        admin::set_max_referrer_bonus_xlm(env, admin, cap)
     }
 
     /// Update the global timeout in seconds.
-    ///
-    /// This acts as the governance/admin override for new escrows.
     pub fn set_default_timeout_seconds(env: Env, admin: Address, timeout_seconds: u32) {
-        admin.require_auth();
-
-        let stored_admin: Address = env
-            .storage()
-            .instance()
-            .get(&DataKey::Admin)
-            .expect("Not initialized");
-        if stored_admin != admin {
-            panic!("Only admin can update the timeout");
-        }
-        if timeout_seconds == 0 {
-            panic!("Timeout must be positive");
-        }
-
-        env.storage()
-            .instance()
-            .set(&DataKey::DefaultTimeoutSeconds, &timeout_seconds);
-        env.events()
-            .publish((symbol_short!("timeout"), admin), timeout_seconds);
+        admin::set_default_timeout_seconds(env, admin, timeout_seconds)
     }
 
-    /// Admin freezes the entire contract — all state-mutating operations are
-    /// blocked until unfreeze_contract() is called with enough admin signatures.
-    ///
-    /// Any admin in the stored admin list may call this function.
+    /// Admin freezes the entire contract.
     pub fn freeze_contract(env: Env, admin: Address) {
-        admin.require_auth();
-
-        let admins: Vec<Address> = env
-            .storage()
-            .instance()
-            .get(&DataKey::Admins)
-            .expect("Not initialized");
-
-        if !admins.contains(&admin) {
-            panic!("Only an admin can freeze the contract");
-        }
-
-        env.storage().instance().set(&DataKey::Frozen, &true);
-
-        env.events().publish((symbol_short!("frozen"), admin), true);
+        admin::freeze_contract(env, admin)
     }
 
     /// Unfreeze the contract — requires M-of-N admin signatures.
-    ///
-    /// `admins` must contain at least `UnfreezeThreshold` distinct admin
-    /// addresses, each of which must also authorize the call via `require_auth`.
-    /// The addresses in `admins` must all be present in the stored admin list.
     pub fn unfreeze_contract(env: Env, admins: Vec<Address>) {
-        let threshold: u32 = env
-            .storage()
-            .instance()
-            .get(&DataKey::UnfreezeThreshold)
-            .expect("Not initialized");
-
-        if admins.len() < threshold {
-            panic!("Insufficient admin signatures to unfreeze");
-        }
-
-        let stored_admins: Vec<Address> = env
-            .storage()
-            .instance()
-            .get(&DataKey::Admins)
-            .expect("Not initialized");
-
-        for admin in admins.iter() {
-            admin.require_auth();
-            if !stored_admins.contains(&admin) {
-                panic!("One of the provided addresses is not an admin");
-            }
-        }
-
-        // De-duplication guard: every admin in `admins` must be distinct.
-        let mut seen: Vec<Address> = Vec::new(&env);
-        for admin in admins.iter() {
-            if seen.contains(&admin) {
-                panic!("Duplicate admin in unfreeze signatures");
-            }
-            seen.push_back(admin);
-        }
-
-        let was_frozen: bool = env
-            .storage()
-            .instance()
-            .get(&DataKey::Frozen)
-            .unwrap_or(false);
-
-        if !was_frozen {
-            panic!("Contract is not frozen");
-        }
-
-        env.storage().instance().set(&DataKey::Frozen, &false);
-
-        env.events()
-            .publish((symbol_short!("unfroz"), threshold), admins.len());
+        admin::unfreeze_contract(env, admins)
     }
 
     /// Add a new admin address to the multi-sig admin list.
-    /// Requires auth from an existing admin.
     pub fn add_admin(env: Env, admin: Address, new_admin: Address) {
-        admin.require_auth();
-
-        let mut admins: Vec<Address> = env
-            .storage()
-            .instance()
-            .get(&DataKey::Admins)
-            .expect("Not initialized");
-
-        if !admins.contains(&admin) {
-            panic!("Only an admin can add new admins");
-        }
-        if admins.contains(&new_admin) {
-            panic!("Address is already an admin");
-        }
-
-        admins.push_back(new_admin);
-        env.storage().instance().set(&DataKey::Admins, &admins);
+        admin::add_admin(env, admin, new_admin)
     }
 
     /// Update the unfreeze threshold (the M in M-of-N).
-    /// Requires auth from an existing admin.
     pub fn set_unfreeze_threshold(env: Env, admin: Address, threshold: u32) {
-        admin.require_auth();
-
-        let admins: Vec<Address> = env
-            .storage()
-            .instance()
-            .get(&DataKey::Admins)
-            .expect("Not initialized");
-
-        if !admins.contains(&admin) {
-            panic!("Only an admin can update the threshold");
-        }
-        if threshold == 0 || threshold > admins.len() {
-            panic!("Threshold must be between 1 and the number of admins");
-        }
-
-        env.storage()
-            .instance()
-            .set(&DataKey::UnfreezeThreshold, &threshold);
+        admin::set_unfreeze_threshold(env, admin, threshold)
     }
 
     /// Return the list of admin addresses.
     pub fn get_admins(env: Env) -> Vec<Address> {
-        env.storage()
-            .instance()
-            .get(&DataKey::Admins)
-            .expect("Not initialized")
+        admin::get_admins(env)
     }
 
     /// Return the unfreeze threshold.
     pub fn get_unfreeze_threshold(env: Env) -> u32 {
-        env.storage()
-            .instance()
-            .get(&DataKey::UnfreezeThreshold)
-            .unwrap_or(2)
+        admin::get_unfreeze_threshold(env)
     }
 
     // ─── Escrow Timeout Extension by Mutual Consent ────────────────────────
 
     /// Either party may request to extend the escrow timeout.
-    /// Stores a pending ExtensionRequest for the other party to approve.
     pub fn request_extension(env: Env, job_id: String, caller: Address, new_timeout_ledger: u32) {
-        caller.require_auth();
-
-        let escrow: Escrow = env
-            .storage()
-            .instance()
-            .get(&DataKey::Escrow(job_id.clone()))
-            .expect("Escrow not found");
-
-        if caller != escrow.client && caller != escrow.freelancer {
-            panic!("Only the client or freelancer can request an extension");
-        }
-        if escrow.status != EscrowStatus::Locked && escrow.status != EscrowStatus::InProgress {
-            panic!("Cannot extend timeout in current status");
-        }
-        if new_timeout_ledger <= escrow.timeout_ledger {
-            panic!("New timeout must be later than current timeout");
-        }
-        if env
-            .storage()
-            .instance()
-            .has(&DataKey::ExtensionRequest(job_id.clone()))
-        {
-            panic!("An extension request is already pending for this job");
-        }
-
-        let request = ExtensionRequest {
-            requested_by: caller.clone(),
-            new_timeout_ledger,
-            created_at: env.ledger().sequence(),
-        };
-
-        env.storage()
-            .instance()
-            .set(&DataKey::ExtensionRequest(job_id.clone()), &request);
-
-        env.events().publish(
-            (symbol_short!("ext_req"), job_id.clone()),
-            (caller, new_timeout_ledger),
-        );
+        escrow::request_extension(env, job_id, caller, new_timeout_ledger)
     }
 
-    /// The other party approves the pending extension, updating the escrow's
-    /// timeout_ledger and TimeoutTimestamp atomically.
+    /// The other party approves the pending extension.
     pub fn approve_extension(env: Env, job_id: String, caller: Address) {
-        caller.require_auth();
-
-        let mut escrow: Escrow = env
-            .storage()
-            .instance()
-            .get(&DataKey::Escrow(job_id.clone()))
-            .expect("Escrow not found");
-
-        if escrow.status != EscrowStatus::Locked && escrow.status != EscrowStatus::InProgress {
-            panic!("Cannot extend timeout in current status");
-        }
-
-        let request: ExtensionRequest = env
-            .storage()
-            .instance()
-            .get(&DataKey::ExtensionRequest(job_id.clone()))
-            .expect("No pending extension request");
-
-        if caller == request.requested_by {
-            panic!("Cannot approve your own extension request");
-        }
-        if caller != escrow.client && caller != escrow.freelancer {
-            panic!("Only the client or freelancer can approve an extension");
-        }
-
-        let ledger_delta = request
-            .new_timeout_ledger
-            .checked_sub(escrow.timeout_ledger)
-            .expect("Arithmetic underflow");
-
-        escrow.timeout_ledger = request.new_timeout_ledger;
-        env.storage()
-            .instance()
-            .set(&DataKey::Escrow(job_id.clone()), &escrow);
-
-        let current_timestamp = env.ledger().timestamp() as u32;
-        let timeout_timestamp: u32 = env
-            .storage()
-            .instance()
-            .get(&DataKey::TimeoutTimestamp(job_id.clone()))
-            .unwrap_or(current_timestamp);
-        let approx_seconds_per_ledger: u32 = 5;
-        let timestamp_extension = ledger_delta
-            .checked_mul(approx_seconds_per_ledger)
-            .expect("Arithmetic overflow");
-        let new_timeout_timestamp = timeout_timestamp
-            .checked_add(timestamp_extension)
-            .expect("Timestamp overflow");
-        env.storage().instance().set(
-            &DataKey::TimeoutTimestamp(job_id.clone()),
-            &new_timeout_timestamp,
-        );
-
-        env.storage()
-            .instance()
-            .remove(&DataKey::ExtensionRequest(job_id.clone()));
-
-        env.events().publish(
-            (symbol_short!("ext_app"), job_id.clone()),
-            (caller, request.requested_by, request.new_timeout_ledger),
-        );
+        escrow::approve_extension(env, job_id, caller)
     }
 
     /// Return the pending extension request for a job, if any.
-    pub fn get_extension_request(env: Env, job_id: String) -> Option<ExtensionRequest> {
-        env.storage()
-            .instance()
-            .get(&DataKey::ExtensionRequest(job_id))
+    pub fn get_extension_request(env: Env, job_id: String) -> Option<types::ExtensionRequest> {
+        escrow::get_extension_request(env, job_id)
     }
 
     // ─── On-chain Message Notarization ─────────────────────────────────────
-    //
-    // Messages are stored off-chain on IPFS.  Only the IPFS CID is stored on-chain
-    // via events, providing censorship resistance and verifiability without the
-    // cost of storing full message content on-chain.
 
     /// Publish a message CID to the ledger.
-    ///
-    /// The message content itself is stored off-chain (IPFS).  This function
-    /// records the IPFS CID on-chain so recipients can verify message authenticity
-    /// from Stellar Explorer.
-    ///
-    /// Parameters:
-    ///   job_id    — job this message belongs to
-    ///   sender    — the party sending the message
-    ///   recipient — the party receiving the message
-    ///   ipfs_cid  — IPFS content identifier for the encrypted message payload
     pub fn publish_message(
         env: Env,
         job_id: String,
@@ -1623,39 +304,12 @@ impl MarketPayContract {
         recipient: Address,
         ipfs_cid: String,
     ) {
-        sender.require_auth();
-        Self::check_not_frozen(&env);
-
-        // Basic validation
-        if ipfs_cid.is_empty() {
-            panic!("IPFS CID cannot be empty");
-        }
-
-        // Store CID in contract storage for on-chain verification
-        let mut cids: soroban_sdk::Vec<String> = env
-            .storage()
-            .instance()
-            .get(&DataKey::MessageCid(job_id.clone()))
-            .unwrap_or_else(|| soroban_sdk::Vec::new(&env));
-        cids.push_back(ipfs_cid.clone());
-        env.storage()
-            .instance()
-            .set(&DataKey::MessageCid(job_id.clone()), &cids);
-
-        let ledger_seq = env.ledger().sequence();
-
-        env.events().publish(
-            (symbol_short!("msg_sent"), job_id.clone()),
-            (sender.clone(), recipient.clone(), ipfs_cid, ledger_seq),
-        );
+        messages::publish_message(env, job_id, sender, recipient, ipfs_cid)
     }
 
     /// Retrieve all message CIDs stored on-chain for a job.
     pub fn get_message_cids(env: Env, job_id: String) -> soroban_sdk::Vec<String> {
-        env.storage()
-            .instance()
-            .get(&DataKey::MessageCid(job_id))
-            .unwrap_or_else(|| soroban_sdk::Vec::new(&env))
+        messages::get_message_cids(env, job_id)
     }
 
     // ─── Governance (DAO) ───────────────────────────────────────────────────
@@ -1667,281 +321,33 @@ impl MarketPayContract {
         description: String,
         duration_ledgers: u32,
     ) -> u32 {
-        proposer.require_auth();
-        Self::check_not_frozen(&env);
-
-        if duration_ledgers == 0 {
-            panic!("Duration must be positive");
-        }
-
-        let count: u32 = env
-            .storage()
-            .instance()
-            .get(&DataKey::ProposalCount)
-            .unwrap_or(0);
-        let proposal_id = count.checked_add(1).expect("Counter overflow");
-        let deadline_ledger = env
-            .ledger()
-            .sequence()
-            .checked_add(duration_ledgers)
-            .expect("Arithmetic overflow");
-
-        let proposal = Proposal {
-            id: proposal_id,
-            title: title.clone(),
-            description: description.clone(),
-            votes_for: 0,
-            votes_against: 0,
-            deadline_ledger,
-            resolved: false,
-            result: false,
-        };
-
-        env.storage()
-            .instance()
-            .set(&DataKey::Proposal(proposal_id), &proposal);
-        env.storage()
-            .instance()
-            .set(&DataKey::ProposalCount, &proposal_id);
-
-        env.events().publish(
-            (symbol_short!("proposed"), proposer),
-            (proposal_id, title, deadline_ledger),
-        );
-
-        proposal_id
+        governance::create_proposal(env, proposer, title, description, duration_ledgers)
     }
 
     pub fn cast_vote(env: Env, voter: Address, proposal_id: u32, approve: bool) {
-        voter.require_auth();
-        Self::check_not_frozen(&env);
-
-        let mut proposal: Proposal = env
-            .storage()
-            .instance()
-            .get(&DataKey::Proposal(proposal_id))
-            .expect("Proposal not found");
-
-        if proposal.resolved {
-            panic!("Proposal already resolved");
-        }
-
-        if env.ledger().sequence() >= proposal.deadline_ledger {
-            panic!("Voting period has ended");
-        }
-
-        // Check eligibility: must have completed at least 1 job
-        let jobs: u32 = env
-            .storage()
-            .instance()
-            .get(&DataKey::CompletedJobs(voter.clone()))
-            .unwrap_or(0);
-        if jobs == 0 {
-            panic!("Only users with completed jobs can vote");
-        }
-
-        // Check if already voted
-        let voted_key = DataKey::HasVoted(voter.clone(), proposal_id);
-        if env.storage().instance().has(&voted_key) {
-            panic!("Voter has already cast a vote");
-        }
-
-        if approve {
-            proposal.votes_for = proposal.votes_for.checked_add(1).expect("Counter overflow");
-        } else {
-            proposal.votes_against = proposal
-                .votes_against
-                .checked_add(1)
-                .expect("Counter overflow");
-        }
-
-        env.storage().instance().set(&voted_key, &true);
-        env.storage()
-            .instance()
-            .set(&DataKey::Proposal(proposal_id), &proposal);
-
-        env.events()
-            .publish((symbol_short!("voted"), voter), (proposal_id, approve));
+        governance::cast_vote(env, voter, proposal_id, approve)
     }
 
     pub fn resolve_proposal(env: Env, proposal_id: u32) {
-        Self::check_not_frozen(&env);
-
-        let mut proposal: Proposal = env
-            .storage()
-            .instance()
-            .get(&DataKey::Proposal(proposal_id))
-            .expect("Proposal not found");
-
-        if proposal.resolved {
-            panic!("Proposal already resolved");
-        }
-
-        if env.ledger().sequence() < proposal.deadline_ledger {
-            panic!("Voting period is not over yet");
-        }
-
-        proposal.resolved = true;
-        proposal.result = proposal.votes_for > proposal.votes_against;
-
-        env.storage()
-            .instance()
-            .set(&DataKey::Proposal(proposal_id), &proposal);
-
-        env.events().publish(
-            (symbol_short!("resolved"), proposal_id),
-            (proposal.result, proposal.votes_for, proposal.votes_against),
-        );
+        governance::resolve_proposal(env, proposal_id)
     }
 
-    pub fn get_proposal(env: Env, id: u32) -> Proposal {
-        env.storage()
-            .instance()
-            .get(&DataKey::Proposal(id))
-            .expect("Proposal not found")
+    pub fn get_proposal(env: Env, id: u32) -> types::Proposal {
+        governance::get_proposal(env, id)
     }
 
-    pub fn list_active_proposals(env: Env) -> Vec<Proposal> {
-        let count: u32 = env
-            .storage()
-            .instance()
-            .get(&DataKey::ProposalCount)
-            .unwrap_or(0);
-        let mut active = Vec::new(&env);
-        for id in 1..=count {
-            if let Some(proposal) = env
-                .storage()
-                .instance()
-                .get::<_, Proposal>(&DataKey::Proposal(id))
-            {
-                if !proposal.resolved {
-                    active.push_back(proposal);
-                }
-            }
-        }
-        active
+    pub fn list_active_proposals(env: Env) -> Vec<types::Proposal> {
+        governance::list_active_proposals(env)
     }
 
-    // ─── Placeholders ─────────────────────────────────────────────────────────
+    // ─── Disputes ──────────────────────────────────────────────────────────
 
     /// Raise a dispute — requires admin resolution.
-    ///
-    /// Issue #437: the caller must lock a configurable bond before the
-    /// dispute is accepted.  The bond is enforced only when the admin has
-    /// configured a non-zero bond amount via `set_dispute_bond`; if no
-    /// configuration is present, this function preserves the legacy
-    /// zero-cost behaviour so escrows created before #437 continue to
-    /// function without admin migration.
-    ///
-    /// Soroban's `caller.require_auth()` authorises every token operation
-    /// the contract performs on behalf of the caller within this call,
-    /// so the bond transfer below does NOT need a separate `token.authorize`
-    /// step.
     pub fn raise_dispute(env: Env, job_id: String, caller: Address) {
-        caller.require_auth();
-        Self::check_not_frozen(&env);
-
-        let mut escrow: Escrow = env
-            .storage()
-            .instance()
-            .get(&DataKey::Escrow(job_id.clone()))
-            .expect("Escrow not found");
-
-        if escrow.client != caller && escrow.freelancer != caller {
-            panic!("Only participants can raise a dispute");
-        }
-
-        if escrow.status == EscrowStatus::Released
-            || escrow.status == EscrowStatus::Refunded
-            || escrow.status == EscrowStatus::Frozen
-            || escrow.status == EscrowStatus::Disputed
-        {
-            panic!("Cannot dispute a resolved, frozen, or already-disputed escrow");
-        }
-
-        // Optional bond requirement (Issue #437).  When the admin has not
-        // configured a dispute bond this block is a no-op and the function
-        // falls through to the legacy behaviour preserved for backward
-        // compatibility with pre-#437 escrows and tests.
-        if let Some(bond_cfg) = env
-            .storage()
-            .instance()
-            .get::<_, DisputeBondConfig>(&DataKey::DisputeBondConfig)
-        {
-            // Snapshot the bond into per-job storage FIRST so that an event
-            // consumer / indexer never sees a `bond_lck` event for which
-            // there is no recoverable record.  We update the escrow status
-            // and persist everything before performing the external token
-            // transfer so that storage state is always the truth.
-            env.storage().instance().set(
-                &DataKey::DisputeBond(job_id.clone()),
-                &DisputeBond {
-                    caller: caller.clone(),
-                    token: bond_cfg.token.clone(),
-                    amount: bond_cfg.amount,
-                    raised_at_ledger: env.ledger().sequence(),
-                },
-            );
-
-            escrow.status = EscrowStatus::Disputed;
-            env.storage()
-                .instance()
-                .set(&DataKey::Escrow(job_id.clone()), &escrow);
-
-            // Lock the bond.  `caller.require_auth()` above has already
-            // authorised ALL token operations from this caller, so this
-            // single transfer call covers the bond lock.
-            let bond_token_client = token::Client::new(&env, &bond_cfg.token);
-            let contract_address = env.current_contract_address();
-            bond_token_client.transfer(&caller, &contract_address, &bond_cfg.amount);
-
-            env.events().publish(
-                (symbol_short!("bond_lck"), job_id.clone()),
-                (caller.clone(), bond_cfg.token, bond_cfg.amount),
-            );
-            env.events().publish(
-                (symbol_short!("escrow_ds"), job_id.clone()),
-                (
-                    escrow.client.clone(),
-                    escrow.freelancer.clone(),
-                    caller.clone(),
-                ),
-            );
-            return;
-        }
-
-        // Legacy fallback (zero-cost dispute mode).
-        escrow.status = EscrowStatus::Disputed;
-        env.storage()
-            .instance()
-            .set(&DataKey::Escrow(job_id.clone()), &escrow);
-
-        env.events().publish(
-            (symbol_short!("escrow_ds"), job_id.clone()),
-            (
-                escrow.client.clone(),
-                escrow.freelancer.clone(),
-                caller.clone(),
-            ),
-        );
+        disputes::raise_dispute(env, job_id, caller)
     }
 
-    /// Resolve a disputed escrow with a split-percentage payout and settle
-    /// the dispute bond (Issue #437).
-    ///
-    /// `winner` is the party that prevails in the dispute — they receive
-    /// `split_percentage`% of the escrow amount, and the other party receives
-    /// `(100 - split_percentage)`%.
-    ///
-    /// `winner` must be either the client or the freelancer on the escrow.
-    /// `split_percentage` must be between 0 and 100 inclusive.
-    ///
-    /// The locked dispute bond (if any) is returned to the bond-caller if
-    /// they are the winner, or slashed to the winner if the bond-caller
-    /// is the losing party.
-    ///
-    /// **Arbitrator-only.**  Idempotency is enforced via `DisputeBond`
-    /// storage which is removed after settlement, so a second call panics.
+    /// Resolve a disputed escrow with a split-percentage payout.
     pub fn resolve_dispute(
         env: Env,
         job_id: String,
@@ -1949,406 +355,39 @@ impl MarketPayContract {
         winner: Address,
         split_percentage: u32,
     ) {
-        arbitrator.require_auth();
-        Self::check_not_frozen(&env);
-
-        // ── Only the designated arbitrator may call this function ──────────────
-        let stored_arbitrator: Address = env
-            .storage()
-            .instance()
-            .get(&DataKey::ArbitratorAddress)
-            .expect("No arbitrator configured");
-        if stored_arbitrator != arbitrator {
-            panic!("Only the arbitrator can resolve a dispute");
-        }
-
-        if split_percentage > 100 {
-            panic!("Split percentage must be between 0 and 100");
-        }
-
-        let mut escrow: Escrow = env
-            .storage()
-            .instance()
-            .get(&DataKey::Escrow(job_id.clone()))
-            .expect("Escrow not found");
-
-        if escrow.status != EscrowStatus::Disputed {
-            panic!("Escrow is not in Disputed state");
-        }
-
-        if winner != escrow.client && winner != escrow.freelancer {
-            panic!("Winner must be the client or the freelancer");
-        }
-
-        // Determine loser
-        let loser: Address = if winner == escrow.client {
-            escrow.freelancer.clone()
-        } else {
-            escrow.client.clone()
-        };
-
-        // Calculate split amounts
-        let winner_amount = escrow
-            .amount
-            .checked_mul(split_percentage as i128)
-            .expect("Arithmetic overflow")
-            .checked_div(100)
-            .expect("Arithmetic overflow");
-        let loser_amount = escrow
-            .amount
-            .checked_sub(winner_amount)
-            .expect("Arithmetic underflow");
-
-        // Update escrow status and clean up stale keys BEFORE external transfers
-        escrow.status = EscrowStatus::Released;
-        env.storage()
-            .instance()
-            .set(&DataKey::Escrow(job_id.clone()), &escrow);
-        env.storage()
-            .instance()
-            .remove(&DataKey::TimeoutTimestamp(job_id.clone()));
-        env.storage()
-            .instance()
-            .remove(&DataKey::FreelancerDeliverableHash(job_id.clone()));
-
-        // Pay out the escrow principal — split between winner and loser
-        let escrow_token_client = token::Client::new(&env, &escrow.token);
-        if winner_amount > 0 {
-            escrow_token_client.transfer(&env.current_contract_address(), &winner, &winner_amount);
-        }
-        if loser_amount > 0 {
-            escrow_token_client.transfer(&env.current_contract_address(), &loser, &loser_amount);
-        }
-
-        // Pull snapshot of the locked bond (may be absent if zero-cost mode).
-        let bond: Option<DisputeBond> = env
-            .storage()
-            .instance()
-            .get(&DataKey::DisputeBond(job_id.clone()));
-
-        // Settle the bond — caller-wins returns it, caller-loses slashes it.
-        if let Some(b) = bond.clone() {
-            let bond_token_client = token::Client::new(&env, &b.token);
-            if b.caller == winner {
-                // Bond-caller is the winner → return bond
-                bond_token_client.transfer(&env.current_contract_address(), &b.caller, &b.amount);
-                env.events().publish(
-                    (symbol_short!("bond_rtn"), job_id.clone()),
-                    (b.caller.clone(), b.amount),
-                );
-            } else {
-                // Bond-caller lost → slash bond to winner
-                bond_token_client.transfer(&env.current_contract_address(), &winner, &b.amount);
-                env.events().publish(
-                    (symbol_short!("bond_slsh"), job_id.clone()),
-                    (winner.clone(), b.amount),
-                );
-            }
-            // Consume the bond record so a second resolve_dispute panics
-            env.storage()
-                .instance()
-                .remove(&DataKey::DisputeBond(job_id.clone()));
-        }
-
-        // Emit DisputeResolved event
-        env.events().publish(
-            (symbol_short!("dsp_res"), job_id.clone()),
-            (
-                arbitrator.clone(),
-                winner.clone(),
-                loser.clone(),
-                winner_amount,
-                loser_amount,
-            ),
-        );
+        disputes::resolve_dispute(env, job_id, arbitrator, winner, split_percentage)
     }
 
-    /// Admin sets the global dispute bond configuration (Issue #437).
-    ///
-    /// `amount == 0` and an `Option::None` (key absent) both leave the
-    /// contract in **legacy zero-cost mode** so existing escrows and tests
-    /// continue to operate without modification.  Setting a positive amount
-    /// enables the bond requirement for all SUBSEQUENT disputes (existing
-    /// disputes are unaffected — bonds are snapshotted at lock time).
+    /// Admin sets the global dispute bond configuration.
     pub fn set_dispute_bond(env: Env, admin: Address, token: Address, amount: i128) {
-        admin.require_auth();
-
-        let stored_admin: Address = env
-            .storage()
-            .instance()
-            .get(&DataKey::Admin)
-            .expect("Not initialized");
-        if stored_admin != admin {
-            panic!("Only admin can update the dispute bond");
-        }
-        if amount <= 0 {
-            panic!("Bond amount must be positive");
-        }
-
-        env.storage().instance().set(
-            &DataKey::DisputeBondConfig,
-            &DisputeBondConfig {
-                token: token.clone(),
-                amount,
-            },
-        );
-
-        env.events()
-            .publish((symbol_short!("bond_cfg"), admin), (token, amount));
+        disputes::set_dispute_bond(env, admin, token, amount)
     }
 
-    /// Read the global dispute bond configuration.  Returns `(None, 0)` in
-    /// legacy zero-cost mode (key absent).
+    /// Read the global dispute bond configuration.
     pub fn get_dispute_bond_config(env: Env) -> (Option<Address>, i128) {
-        env.storage()
-            .instance()
-            .get::<_, DisputeBondConfig>(&DataKey::DisputeBondConfig)
-            .map(|c| (Some(c.token), c.amount))
-            .unwrap_or((None, 0))
+        disputes::get_dispute_bond_config(env)
     }
 
-    /// Read the per-job locked bond record.  Returns `None` if no bond
-    /// was locked (either legacy zero-cost mode or already settled).
-    pub fn get_dispute_bond(env: Env, job_id: String) -> Option<DisputeBond> {
-        env.storage().instance().get(&DataKey::DisputeBond(job_id))
+    /// Read the per-job locked bond record.
+    pub fn get_dispute_bond(env: Env, job_id: String) -> Option<types::DisputeBond> {
+        disputes::get_dispute_bond(env, job_id)
     }
+
+    // ─── Milestones ────────────────────────────────────────────────────────
 
     /// Milestone-based partial release.
-    /// Can be called even if the escrow is Disputed, to release completed work.
     pub fn release_milestone(env: Env, job_id: String, milestone_id: u32, client: Address) {
-        client.require_auth();
-        Self::check_not_frozen(&env);
-
-        let mut escrow: Escrow = env
-            .storage()
-            .instance()
-            .get(&DataKey::Escrow(job_id.clone()))
-            .expect("Escrow not found");
-
-        if escrow.client != client {
-            panic!("Only the client can release a milestone");
-        }
-        if escrow.status != EscrowStatus::InProgress
-            && escrow.status != EscrowStatus::Locked
-            && escrow.status != EscrowStatus::Disputed
-        {
-            panic!("Cannot release milestone in current status");
-        }
-
-        let mut idx: Option<u32> = None;
-        for i in 0..escrow.milestones.len() {
-            if escrow.milestones.get(i).unwrap().id == milestone_id {
-                idx = Some(i);
-                break;
-            }
-        }
-        let milestone_index = idx.expect("Invalid milestone id");
-
-        let mut milestone = escrow.milestones.get(milestone_index).unwrap();
-        if milestone.released {
-            panic!("Milestone already released");
-        }
-        if milestone.rejected {
-            panic!("Milestone already rejected");
-        }
-
-        milestone.released = true;
-        escrow.milestones.set(milestone_index, milestone.clone());
-
-        // Compute payout for this milestone's percentage of the total
-        let payout = escrow
-            .amount
-            .checked_mul(milestone.percentage as i128)
-            .expect("Arithmetic overflow")
-            .checked_div(100)
-            .expect("Arithmetic overflow");
-
-        let token_client = token::Client::new(&env, &escrow.token);
-
-        // ── Platform fee ────────────────────────────────────────────────────
-        let fee_bps: u32 = env
-            .storage()
-            .instance()
-            .get(&DataKey::PlatformFeeBps)
-            .unwrap_or(0);
-        let treasury: Address = env
-            .storage()
-            .instance()
-            .get(&DataKey::TreasuryAddress)
-            .expect("Treasury not set");
-        let fee_amount = payout
-            .checked_mul(fee_bps as i128)
-            .expect("Arithmetic overflow")
-            .checked_div(10_000)
-            .expect("Arithmetic overflow");
-        let to_freelancer = payout.checked_sub(fee_amount).expect("Arithmetic overflow");
-
-        if fee_amount > 0 {
-            token_client.transfer(&env.current_contract_address(), &treasury, &fee_amount);
-            env.events().publish(
-                (symbol_short!("plat_fee"), job_id.clone()),
-                (treasury.clone(), fee_amount),
-            );
-        }
-
-        // Transfer remaining funds to freelancer
-        token_client.transfer(
-            &env.current_contract_address(),
-            &escrow.freelancer,
-            &to_freelancer,
-        );
-
-        // Check if all milestones are now resolved (released or rejected)
-        let mut all_completed = true;
-        for ms in escrow.milestones.iter() {
-            if !ms.released && !ms.rejected {
-                all_completed = false;
-                break;
-            }
-        }
-
-        if all_completed {
-            escrow.status = EscrowStatus::Released;
-            env.storage()
-                .instance()
-                .remove(&DataKey::TimeoutTimestamp(job_id.clone()));
-
-            // Increment CompletedJobs for the freelancer and client
-            let freelancer_jobs: u32 = env
-                .storage()
-                .instance()
-                .get(&DataKey::CompletedJobs(escrow.freelancer.clone()))
-                .unwrap_or(0);
-            let new_freelancer_jobs = freelancer_jobs.checked_add(1).expect("Counter overflow");
-            env.storage().instance().set(
-                &DataKey::CompletedJobs(escrow.freelancer.clone()),
-                &new_freelancer_jobs,
-            );
-
-            let client_jobs: u32 = env
-                .storage()
-                .instance()
-                .get(&DataKey::CompletedJobs(escrow.client.clone()))
-                .unwrap_or(0);
-            let new_client_jobs = client_jobs.checked_add(1).expect("Counter overflow");
-            env.storage().instance().set(
-                &DataKey::CompletedJobs(escrow.client.clone()),
-                &new_client_jobs,
-            );
-        }
-
-        env.storage()
-            .instance()
-            .set(&DataKey::Escrow(job_id.clone()), &escrow);
-
-        env.events().publish(
-            (Symbol::new(&env, "milestone_released"), job_id.clone()),
-            (
-                escrow.client.clone(),
-                escrow.freelancer.clone(),
-                milestone_id,
-                payout,
-            ),
-        );
+        milestones::release_milestone(env, job_id, milestone_id, client)
     }
 
-    /// Partial milestone refund — the client rejects a single milestone and its
-    /// share of the escrow is returned to the client. Remaining milestones stay
-    /// locked in the contract.
-    ///
-    /// Only the client may call this. The milestone is identified by its id
-    /// (the index assigned at creation time).
+    /// Partial milestone refund — the client rejects a single milestone.
     pub fn reject_milestone(env: Env, job_id: String, milestone_index: u32, client: Address) {
-        client.require_auth();
-        Self::check_not_frozen(&env);
-
-        let mut escrow: Escrow = env
-            .storage()
-            .instance()
-            .get(&DataKey::Escrow(job_id.clone()))
-            .expect("Escrow not found");
-
-        if escrow.client != client {
-            panic!("Only the client can reject a milestone");
-        }
-        if escrow.status != EscrowStatus::InProgress
-            && escrow.status != EscrowStatus::Locked
-            && escrow.status != EscrowStatus::Disputed
-        {
-            panic!("Cannot reject milestone in current status");
-        }
-
-        let mut idx: Option<u32> = None;
-        for i in 0..escrow.milestones.len() {
-            if escrow.milestones.get(i).unwrap().id == milestone_index {
-                idx = Some(i);
-                break;
-            }
-        }
-        let position = idx.expect("Invalid milestone id");
-
-        let mut milestone = escrow.milestones.get(position).unwrap();
-        if milestone.released {
-            panic!("Milestone already released");
-        }
-        if milestone.rejected {
-            panic!("Milestone already rejected");
-        }
-
-        milestone.rejected = true;
-        escrow.milestones.set(position, milestone.clone());
-
-        // Compute this milestone's percentage of the total and refund to client
-        let refund = escrow
-            .amount
-            .checked_mul(milestone.percentage as i128)
-            .expect("Arithmetic overflow")
-            .checked_div(100)
-            .expect("Arithmetic overflow");
-
-        let token_client = token::Client::new(&env, &escrow.token);
-        token_client.transfer(&env.current_contract_address(), &escrow.client, &refund);
-
-        // If every milestone is now resolved (released or rejected), close out the escrow
-        let mut all_resolved = true;
-        for ms in escrow.milestones.iter() {
-            if !ms.released && !ms.rejected {
-                all_resolved = false;
-                break;
-            }
-        }
-        if all_resolved {
-            escrow.status = EscrowStatus::Released;
-            env.storage()
-                .instance()
-                .remove(&DataKey::TimeoutTimestamp(job_id.clone()));
-        }
-
-        env.storage()
-            .instance()
-            .set(&DataKey::Escrow(job_id.clone()), &escrow);
-
-        env.events().publish(
-            (Symbol::new(&env, "milestone_rejected"), job_id.clone()),
-            (
-                escrow.client.clone(),
-                escrow.freelancer.clone(),
-                milestone_index,
-                refund,
-            ),
-        );
+        milestones::reject_milestone(env, job_id, milestone_index, client)
     }
 
-    // ─── Issue #344: Job Boost with XLM Payment ──────────────────────────────
+    // ─── Job Boost ─────────────────────────────────────────────────────────
 
     /// Client pays XLM to the platform treasury to boost a job listing.
-    ///
-    /// Boost tiers (in stroops, 1 XLM = 10_000_000 stroops):
-    ///   ≥  5 XLM → 7-day boost
-    ///   ≥ 15 XLM → 30-day boost
-    ///
-    /// The payment is transferred directly to `treasury`.
-    /// Emits a `JobBoosted` event with job_id and boost_expiry_ledger.
     pub fn boost_job(
         env: Env,
         job_id: String,
@@ -2357,108 +396,27 @@ impl MarketPayContract {
         token: Address,
         amount: i128,
     ) {
-        client.require_auth();
-        Self::check_not_frozen(&env);
-
-        if amount <= 0 {
-            panic!("Boost amount must be positive");
-        }
-
-        // Minimum boost is 5 XLM (50_000_000 stroops)
-        let min_boost_stroops: i128 = 50_000_000;
-        if amount < min_boost_stroops {
-            panic!("Minimum boost is 5 XLM");
-        }
-
-        // Transfer payment from client to treasury
-        let token_client = token::Client::new(&env, &token);
-        token_client.transfer(&client, &treasury, &amount);
-
-        // Calculate boost duration in ledgers (~5 s/ledger)
-        // 7 days  = 120_960 ledgers
-        // 30 days = 518_400 ledgers
-        let boost_ledgers: u32 = if amount >= 150_000_000 {
-            518_400 // 30 days
-        } else {
-            120_960 // 7 days
-        };
-
-        let boost_expiry = env
-            .ledger()
-            .sequence()
-            .checked_add(boost_ledgers)
-            .expect("Boost expiry overflow");
-
-        env.events().publish(
-            (symbol_short!("boosted"), client),
-            (job_id, boost_expiry, amount),
-        );
+        escrow::boost_job(env, job_id, client, treasury, token, amount)
     }
 
-    // ─── Issue #108: Sealed-Bid Budget Commitment ────────────────────────────
+    // ─── Sealed-Bid Budget Commitment ──────────────────────────────────────
 
-    /// Client commits to a budget amount (sealed-bid, prevents anchoring bias).
+    /// Client commits to a budget amount (sealed-bid).
     pub fn commit_budget(env: Env, job_id: String, budget_amount: i128, client: Address) {
-        client.require_auth();
-        Self::check_not_frozen(&env);
-
-        if budget_amount <= 0 {
-            panic!("Budget must be positive");
-        }
-
-        let commitment = BudgetCommitment {
-            job_id: job_id.clone(),
-            client: client.clone(),
-            budget_amount,
-            is_revealed: false,
-        };
-
-        env.storage()
-            .instance()
-            .set(&DataKey::BudgetCommitment(job_id.clone()), &commitment);
-
-        env.events()
-            .publish((symbol_short!("budgtcmt"), client), job_id);
+        auction::commit_budget(env, job_id, budget_amount, client)
     }
 
     /// Reveal the budget. Auto-rejects bids over 150% of budget.
     pub fn reveal_budget(env: Env, job_id: String, client: Address) {
-        client.require_auth();
-        Self::check_not_frozen(&env);
-
-        let mut commitment: BudgetCommitment = env
-            .storage()
-            .instance()
-            .get(&DataKey::BudgetCommitment(job_id.clone()))
-            .expect("Budget commitment not found");
-
-        if commitment.client != client {
-            panic!("Only the client can reveal the budget");
-        }
-        if commitment.is_revealed {
-            panic!("Budget already revealed");
-        }
-
-        commitment.is_revealed = true;
-        env.storage()
-            .instance()
-            .set(&DataKey::BudgetCommitment(job_id.clone()), &commitment);
-
-        env.events().publish(
-            (symbol_short!("budgrvld"), client),
-            commitment.budget_amount,
-        );
+        auction::reveal_budget(env, job_id, client)
     }
 
     /// Get budget commitment.
-    pub fn get_budget_commitment(env: Env, job_id: String) -> BudgetCommitment {
-        env.storage()
-            .instance()
-            .get(&DataKey::BudgetCommitment(job_id))
-            .expect("Budget commitment not found")
+    pub fn get_budget_commitment(env: Env, job_id: String) -> types::BudgetCommitment {
+        auction::get_budget_commitment(env, job_id)
     }
 
-    // ─── Issue #338: Sealed-Bid Commitment Scheme ───────────────────────────
+    // ─── Sealed-Bid Commitment Scheme ──────────────────────────────────────
 
     /// Freelancer submits a sealed commitment hash for their bid amount.
     pub fn submit_bid_commitment(
@@ -2467,86 +425,12 @@ impl MarketPayContract {
         freelancer: Address,
         commitment: BytesN<32>,
     ) {
-        freelancer.require_auth();
-        Self::check_not_frozen(&env);
-
-        // Ensure this job has a client-owned bidding session via budget commitment.
-        let _budget: BudgetCommitment = env
-            .storage()
-            .instance()
-            .get(&DataKey::BudgetCommitment(job_id.clone()))
-            .expect("Budget commitment not found");
-
-        if let Some(state) = env
-            .storage()
-            .instance()
-            .get::<_, BiddingState>(&DataKey::BiddingState(job_id.clone()))
-        {
-            if state.is_closed {
-                panic!("Bidding is closed");
-            }
-        }
-
-        let key = DataKey::BidCommitment(job_id.clone(), freelancer.clone());
-        if env.storage().instance().has(&key) {
-            panic!("Bid commitment already submitted");
-        }
-
-        let bid_commitment = BidCommitment {
-            job_id: job_id.clone(),
-            freelancer: freelancer.clone(),
-            commitment,
-            submitted_at_ledger: env.ledger().sequence(),
-            bid_revealed: false,
-        };
-
-        env.storage().instance().set(&key, &bid_commitment);
-        env.events()
-            .publish((symbol_short!("bid_cmt"), job_id), freelancer);
+        auction::submit_bid_commitment(env, job_id, freelancer, commitment)
     }
 
     /// Client closes bidding and opens a reveal window.
     pub fn close_bidding(env: Env, job_id: String, client: Address) {
-        client.require_auth();
-        Self::check_not_frozen(&env);
-
-        let budget: BudgetCommitment = env
-            .storage()
-            .instance()
-            .get(&DataKey::BudgetCommitment(job_id.clone()))
-            .expect("Budget commitment not found");
-        if budget.client != client {
-            panic!("Only the client can close bidding");
-        }
-
-        if let Some(existing) = env
-            .storage()
-            .instance()
-            .get::<_, BiddingState>(&DataKey::BiddingState(job_id.clone()))
-        {
-            if existing.is_closed {
-                panic!("Bidding already closed");
-            }
-        }
-
-        let closed_at = env.ledger().sequence();
-        let reveal_deadline = closed_at
-            .checked_add(REVEAL_WINDOW_LEDGERS)
-            .expect("Reveal deadline overflow");
-
-        let state = BiddingState {
-            job_id: job_id.clone(),
-            client: client.clone(),
-            is_closed: true,
-            closed_at_ledger: closed_at,
-            reveal_deadline_ledger: reveal_deadline,
-        };
-
-        env.storage()
-            .instance()
-            .set(&DataKey::BiddingState(job_id.clone()), &state);
-        env.events()
-            .publish((symbol_short!("bid_cls"), job_id), reveal_deadline);
+        auction::close_bidding(env, job_id, client)
     }
 
     /// Freelancer reveals their sealed bid: amount + nonce.
@@ -2557,2403 +441,109 @@ impl MarketPayContract {
         amount: i128,
         nonce: BytesN<32>,
     ) {
-        freelancer.require_auth();
-        Self::check_not_frozen(&env);
-
-        if amount <= 0 {
-            panic!("Bid amount must be positive");
-        }
-
-        let state: BiddingState = env
-            .storage()
-            .instance()
-            .get(&DataKey::BiddingState(job_id.clone()))
-            .expect("Bidding not closed");
-        if !state.is_closed {
-            panic!("Bidding not closed");
-        }
-        if env.ledger().sequence() > state.reveal_deadline_ledger {
-            panic!("Reveal window has closed");
-        }
-
-        let key = DataKey::BidCommitment(job_id.clone(), freelancer.clone());
-        let mut bid_commitment: BidCommitment = env
-            .storage()
-            .instance()
-            .get(&key)
-            .expect("Bid commitment not found");
-
-        if bid_commitment.bid_revealed {
-            panic!("Bid already revealed");
-        }
-
-        let expected = Self::compute_bid_commitment(&env, amount, nonce);
-        if expected != bid_commitment.commitment {
-            panic!("Commitment verification failed");
-        }
-
-        bid_commitment.bid_revealed = true;
-        env.storage().instance().set(&key, &bid_commitment);
-
-        let mut reveals: Vec<RevealedBid> = env
-            .storage()
-            .instance()
-            .get(&DataKey::RevealedBids(job_id.clone()))
-            .unwrap_or_else(|| Vec::new(&env));
-        reveals.push_back(RevealedBid {
-            freelancer: freelancer.clone(),
-            amount,
-            revealed_at_ledger: env.ledger().sequence(),
-        });
-        env.storage()
-            .instance()
-            .set(&DataKey::RevealedBids(job_id.clone()), &reveals);
-
-        env.events()
-            .publish((symbol_short!("bid_rvl"), job_id), (freelancer, amount));
+        auction::reveal_bid(env, job_id, freelancer, amount, nonce)
     }
 
     /// Read a freelancer's sealed bid commitment.
-    pub fn get_bid_commitment(env: Env, job_id: String, freelancer: Address) -> BidCommitment {
-        env.storage()
-            .instance()
-            .get(&DataKey::BidCommitment(job_id, freelancer))
-            .expect("Bid commitment not found")
+    pub fn get_bid_commitment(
+        env: Env,
+        job_id: String,
+        freelancer: Address,
+    ) -> types::BidCommitment {
+        auction::get_bid_commitment(env, job_id, freelancer)
     }
 
     /// Read all bids that were revealed during reveal phase.
-    pub fn get_revealed_bids(env: Env, job_id: String) -> Vec<RevealedBid> {
-        env.storage()
-            .instance()
-            .get(&DataKey::RevealedBids(job_id))
-            .unwrap_or_else(|| Vec::new(&env))
+    pub fn get_revealed_bids(env: Env, job_id: String) -> Vec<types::RevealedBid> {
+        auction::get_revealed_bids(env, job_id)
     }
 
-    // ─── Issue #105: Deliverable Hash Oracle ────────────────────────────────
+    // ─── Deliverable Hash Oracle ───────────────────────────────────────────
 
     /// Client submits deliverable hash.
     pub fn submit_client_deliverable(env: Env, job_id: String, client: Address) {
-        client.require_auth();
-        Self::check_not_frozen(&env);
-
-        let mut submission: DeliverableSubmission = env
-            .storage()
-            .instance()
-            .get(&DataKey::DeliverableSubmission(job_id.clone()))
-            .unwrap_or_else(|| DeliverableSubmission {
-                job_id: job_id.clone(),
-                client_hash_submitted: false,
-                freelancer_hash_submitted: false,
-                hashes_match: false,
-            });
-
-        submission.client_hash_submitted = true;
-        env.storage()
-            .instance()
-            .set(&DataKey::DeliverableSubmission(job_id.clone()), &submission);
-
-        env.events()
-            .publish((symbol_short!("clthash"), client), job_id);
+        deliverable::submit_client_deliverable(env, job_id, client)
     }
 
     /// Freelancer submits deliverable hash.
     pub fn submit_freelancer_deliverable(env: Env, job_id: String, freelancer: Address) {
-        freelancer.require_auth();
-        Self::check_not_frozen(&env);
-
-        let mut submission: DeliverableSubmission = env
-            .storage()
-            .instance()
-            .get(&DataKey::DeliverableSubmission(job_id.clone()))
-            .unwrap_or_else(|| DeliverableSubmission {
-                job_id: job_id.clone(),
-                client_hash_submitted: false,
-                freelancer_hash_submitted: false,
-                hashes_match: false,
-            });
-
-        submission.freelancer_hash_submitted = true;
-        env.storage()
-            .instance()
-            .set(&DataKey::DeliverableSubmission(job_id.clone()), &submission);
-
-        env.events()
-            .publish((symbol_short!("frelhash"), freelancer), job_id);
+        deliverable::submit_freelancer_deliverable(env, job_id, freelancer)
     }
 
     /// Oracle/freelancer submits the deliverable hash.
-    ///
-    /// If it matches the expected deliverable hash stored in escrow,
-    /// the escrow is auto-released. If mismatched, escrow enters dispute.
     pub fn submit_deliverable(env: Env, job_id: String, actual_hash: BytesN<32>, caller: Address) {
-        caller.require_auth();
-        Self::check_not_frozen(&env);
-
-        let mut escrow: Escrow = env
-            .storage()
-            .instance()
-            .get(&DataKey::Escrow(job_id.clone()))
-            .expect("Escrow not found");
-
-        let admin: Address = env
-            .storage()
-            .instance()
-            .get(&DataKey::Admin)
-            .expect("Not initialized");
-
-        if caller != escrow.freelancer && caller != admin {
-            panic!("Only freelancer or oracle can submit deliverable");
-        }
-
-        let expected_hash = escrow
-            .deliverable_hash
-            .clone()
-            .expect("Escrow has no deliverable hash");
-
-        if actual_hash == expected_hash {
-            // Auto-release on successful deliverable verification.
-            Self::release_escrow_core(env.clone(), job_id.clone(), escrow);
-            env.events()
-                .publish((symbol_short!("dlv_ok"), job_id), (caller, actual_hash));
-            return;
-        }
-
-        // Mismatch must explicitly enter dispute.
-        escrow.status = EscrowStatus::Disputed;
-        env.storage()
-            .instance()
-            .set(&DataKey::Escrow(job_id.clone()), &escrow);
-
-        env.events()
-            .publish((symbol_short!("dlv_bad"), job_id), (caller, actual_hash));
+        deliverable::submit_deliverable(env, job_id, actual_hash, caller)
     }
 
-    /// Auto-release if both hashes match (manual fallback if mismatch after 7 days).
+    /// Auto-release if both hashes match.
     pub fn check_deliverable_match(env: Env, job_id: String) -> bool {
-        Self::check_not_frozen(&env);
-
-        let submission: DeliverableSubmission = env
-            .storage()
-            .instance()
-            .get(&DataKey::DeliverableSubmission(job_id.clone()))
-            .expect("Deliverable submission not found");
-
-        // Both must be submitted
-        if submission.client_hash_submitted && submission.freelancer_hash_submitted {
-            let mut updated = submission.clone();
-            updated.hashes_match = true;
-            env.storage()
-                .instance()
-                .set(&DataKey::DeliverableSubmission(job_id), &updated);
-            return true;
-        }
-        false
+        deliverable::check_deliverable_match(env, job_id)
     }
 
     /// Get deliverable submission status.
-    pub fn get_deliverable_submission(env: Env, job_id: String) -> DeliverableSubmission {
-        env.storage()
-            .instance()
-            .get(&DataKey::DeliverableSubmission(job_id))
-            .expect("Deliverable submission not found")
+    pub fn get_deliverable_submission(env: Env, job_id: String) -> types::DeliverableSubmission {
+        deliverable::get_deliverable_submission(env, job_id)
     }
 
     /// Freelancer submits the SHA-256 hash of the completed deliverable.
-    /// Once submitted, the client can verify and call release_escrow.
     pub fn submit_deliverable_hash(
         env: Env,
         job_id: String,
         freelancer: Address,
         hash: BytesN<32>,
     ) {
-        freelancer.require_auth();
-
-        let escrow: Escrow = env
-            .storage()
-            .instance()
-            .get(&DataKey::Escrow(job_id.clone()))
-            .expect("Escrow not found");
-
-        if escrow.freelancer != freelancer {
-            panic!("Only the freelancer can submit deliverable hash");
-        }
-        if escrow.deliverable_hash.is_none() {
-            panic!("Escrow has no expected deliverable hash");
-        }
-        if escrow.status != EscrowStatus::InProgress && escrow.status != EscrowStatus::Locked {
-            panic!("Can only submit hash for active escrow");
-        }
-
-        env.storage()
-            .instance()
-            .set(&DataKey::FreelancerDeliverableHash(job_id.clone()), &hash);
-
-        env.events()
-            .publish((symbol_short!("dlv_sub"), freelancer), (job_id, hash));
+        deliverable::submit_deliverable_hash(env, job_id, freelancer, hash)
     }
 
     /// Get the freelancer-submitted deliverable hash, if any.
     pub fn get_freelancer_deliverable_hash(env: Env, job_id: String) -> Option<BytesN<32>> {
-        env.storage()
-            .instance()
-            .get(&DataKey::FreelancerDeliverableHash(job_id))
+        deliverable::get_freelancer_deliverable_hash(env, job_id)
     }
 
     /// Verify that the freelancer-submitted hash matches the expected hash.
-    /// Returns true if both exist and match, false otherwise.
     pub fn verify_deliverable_hash(env: Env, job_id: String) -> bool {
-        let escrow: Escrow = env
-            .storage()
-            .instance()
-            .get(&DataKey::Escrow(job_id.clone()))
-            .expect("Escrow not found");
-
-        let Some(expected) = &escrow.deliverable_hash else {
-            return false;
-        };
-
-        let Some(submitted) = env
-            .storage()
-            .instance()
-            .get::<_, BytesN<32>>(&DataKey::FreelancerDeliverableHash(job_id))
-        else {
-            return false;
-        };
-
-        &submitted == expected
+        deliverable::verify_deliverable_hash(env, job_id)
     }
 
-    // ─── Issue #102: Job Completion Certificate ──────────────────────────────
+    // ─── Certificates, Evidence & Ratings ──────────────────────────────────
 
-    /// Mint a certificate when job is completed (upon escrow release).
-    ///
-    /// The certificate is a proof-of-work NFT minted to the freelancer's
-    /// address. `title` is stored on-chain as part of the certificate
-    /// metadata so the certificate carries the job title (not just the id).
+    /// Mint a certificate when job is completed.
     pub fn mint_certificate(env: Env, job_id: String, title: String, client: Address) {
-        client.require_auth();
-        Self::check_not_frozen(&env);
-
-        let escrow: Escrow = env
-            .storage()
-            .instance()
-            .get(&DataKey::Escrow(job_id.clone()))
-            .expect("Escrow not found");
-
-        if client != escrow.client {
-            panic!("Only the escrow client can mint the certificate");
-        }
-        if title.is_empty() {
-            panic!("Certificate title cannot be empty");
-        }
-        if escrow.status != EscrowStatus::Released {
-            panic!("Escrow must be released to mint certificate");
-        }
-        if env
-            .storage()
-            .instance()
-            .has(&DataKey::Certificate(job_id.clone()))
-        {
-            panic!("Certificate already minted");
-        }
-
-        let cert = Certificate {
-            job_id: job_id.clone(),
-            title: title.clone(),
-            client: escrow.client.clone(),
-            freelancer: escrow.freelancer.clone(),
-            amount: escrow.amount,
-            created_at: env.ledger().sequence(),
-        };
-
-        env.storage()
-            .instance()
-            .set(&DataKey::Certificate(job_id.clone()), &cert);
-
-        let mut certs: Vec<String> = env
-            .storage()
-            .instance()
-            .get(&DataKey::FreelancerCertificates(escrow.freelancer.clone()))
-            .unwrap_or_else(|| Vec::new(&env));
-        certs.push_back(job_id.clone());
-        env.storage().instance().set(
-            &DataKey::FreelancerCertificates(escrow.freelancer.clone()),
-            &certs,
-        );
-
-        env.events()
-            .publish((symbol_short!("certmnt"), client), (job_id, escrow.amount));
+        certificates::mint_certificate(env, job_id, title, client)
     }
 
-    /// Append an IPFS CID to a job's on-chain dispute-evidence audit trail
-    /// (Issue #448 --- AC #1).
-    ///
-    /// Caller: the escrow's client OR the escrow's freelancer. The explicit
-    /// `caller` parameter is `require_auth`'d so every chain row carries
-    /// cryptographic provenance of who anchored the CID.
-    ///
-    /// Storage: a Soroban `Vec<Bytes>` of CID bytes is appended at
-    /// `DataKey::EvidenceCids(job_id)`. The vector is append-only; existing
-    /// entries are never overwritten.
+    /// Append an IPFS CID to a job's on-chain dispute-evidence audit trail.
     pub fn submit_evidence_cid(env: Env, job_id: String, cid: Bytes, caller: Address) {
-        caller.require_auth();
-        Self::check_not_frozen(&env);
-
-        if cid.is_empty() {
-            panic!("IPFS CID cannot be empty");
-        }
-
-        let escrow: Escrow = env
-            .storage()
-            .instance()
-            .get(&DataKey::Escrow(job_id.clone()))
-            .expect("Escrow not found");
-
-        if caller != escrow.client && caller != escrow.freelancer {
-            panic!("Only participants can record evidence");
-        }
-
-        if escrow.status == EscrowStatus::Refunded {
-            panic!("Cannot record evidence on a refunded escrow");
-        }
-
-        let mut cids: soroban_sdk::Vec<Bytes> = env
-            .storage()
-            .instance()
-            .get(&DataKey::EvidenceCids(job_id.clone()))
-            .unwrap_or_else(|| soroban_sdk::Vec::new(&env));
-
-        cids.push_back(cid);
-        env.storage()
-            .instance()
-            .set(&DataKey::EvidenceCids(job_id.clone()), &cids);
-
-        env.events().publish(
-            (symbol_short!("evd_add"), job_id),
-            (caller, env.ledger().sequence()),
-        );
+        certificates::submit_evidence_cid(env, job_id, cid, caller)
     }
 
     /// Get a certificate.
-    pub fn get_certificate(env: Env, job_id: String) -> Certificate {
-        env.storage()
-            .instance()
-            .get(&DataKey::Certificate(job_id))
-            .expect("Certificate not found")
+    pub fn get_certificate(env: Env, job_id: String) -> types::Certificate {
+        certificates::get_certificate(env, job_id)
     }
 
     /// Get all certificates for a freelancer.
     pub fn get_freelancer_certificates(env: Env, freelancer: Address) -> Vec<String> {
-        env.storage()
-            .instance()
-            .get(&DataKey::FreelancerCertificates(freelancer))
-            .unwrap_or_else(|| Vec::new(&env))
+        certificates::get_freelancer_certificates(env, freelancer)
     }
 
     pub fn submit_client_rating(env: Env, job_id: String, client: Address, score: u32) {
-        client.require_auth();
-        Self::check_not_frozen(&env);
-        if !(1..=5).contains(&score) {
-            panic!("Score must be between 1 and 5");
-        }
-
-        let escrow: Escrow = env
-            .storage()
-            .instance()
-            .get(&DataKey::Escrow(job_id.clone()))
-            .expect("Escrow not found");
-
-        let mut stats: FreelancerRatingStats = env
-            .storage()
-            .instance()
-            .get(&DataKey::FreelancerRatingStats(escrow.freelancer.clone()))
-            .unwrap_or(FreelancerRatingStats {
-                total_score: 0,
-                count: 0,
-            });
-        stats.total_score = stats
-            .total_score
-            .checked_add(score)
-            .expect("Arithmetic overflow");
-        stats.count = stats.count.checked_add(1).expect("Arithmetic overflow");
-        env.storage()
-            .instance()
-            .set(&DataKey::FreelancerRatingStats(escrow.freelancer), &stats);
+        certificates::submit_client_rating(env, job_id, client, score)
     }
 
     pub fn submit_freelancer_rating(env: Env, job_id: String, freelancer: Address, score: u32) {
-        freelancer.require_auth();
-        Self::check_not_frozen(&env);
-        if !(1..=5).contains(&score) {
-            panic!("Score must be between 1 and 5");
-        }
-
-        let escrow: Escrow = env
-            .storage()
-            .instance()
-            .get(&DataKey::Escrow(job_id.clone()))
-            .expect("Escrow not found");
-        if escrow.status != EscrowStatus::Released {
-            panic!("Ratings are allowed only after escrow release");
-        }
-        if escrow.freelancer != freelancer {
-            panic!("Only job freelancer can submit freelancer rating");
-        }
-        if env
-            .storage()
-            .instance()
-            .has(&DataKey::FreelancerRating(job_id.clone()))
-        {
-            panic!("Freelancer rating already submitted for this job");
-        }
-
-        let rating = Rating {
-            job_id: job_id.clone(),
-            rater: freelancer,
-            rated: escrow.client,
-            score_out_of_5: score,
-            submitted_at_ledger: env.ledger().sequence(),
-        };
-        env.storage()
-            .instance()
-            .set(&DataKey::FreelancerRating(job_id), &rating);
+        certificates::submit_freelancer_rating(env, job_id, freelancer, score)
     }
 
     pub fn resolve_arbitration(env: Env, case_id: u32) {
-        Self::check_not_frozen(&env);
-
-        let mut case: ArbitrationCase = env
-            .storage()
-            .instance()
-            .get(&DataKey::ArbitrationCase(case_id))
-            .expect("Arbitration case not found");
-        if case.votes.len() != 3 {
-            panic!("Exactly 3 votes required");
-        }
-        let vote_a = case.votes.get(0).unwrap();
-        let vote_b = case.votes.get(1).unwrap();
-        let vote_c = case.votes.get(2).unwrap();
-        let min_vote = if vote_a < vote_b { vote_a } else { vote_b };
-        let min_vote = if min_vote < vote_c { min_vote } else { vote_c };
-        let max_vote = if vote_a > vote_b { vote_a } else { vote_b };
-        let max_vote = if max_vote > vote_c { max_vote } else { vote_c };
-        case.resolution = vote_a
-            .checked_add(vote_b)
-            .expect("Counter overflow")
-            .checked_add(vote_c)
-            .expect("Counter overflow")
-            .checked_sub(min_vote)
-            .expect("Arithmetic underflow")
-            .checked_sub(max_vote)
-            .expect("Arithmetic underflow");
-        case.status = 1;
-        env.storage()
-            .instance()
-            .set(&DataKey::ArbitrationCase(case_id), &case);
-
-        env.events()
-            .publish((symbol_short!("arb_rsl"), case_id), case.resolution);
+        certificates::resolve_arbitration(env, case_id)
     }
 
-    /// Read the IPFS CIDs anchoring dispute evidence on-chain for a job
-    /// (Issue #448 --- AC #3). Returns the `Vec<Bytes>` in insertion order
-    /// (oldest first). Empty `Vec` if no evidence has been anchored yet.
+    /// Read the IPFS CIDs anchoring dispute evidence on-chain for a job.
     pub fn get_evidence_cids(env: Env, job_id: String) -> soroban_sdk::Vec<Bytes> {
-        env.storage()
-            .instance()
-            .get(&DataKey::EvidenceCids(job_id))
-            .unwrap_or_else(|| soroban_sdk::Vec::new(&env))
-    }
-}
-
-// ─── Tests ────────────────────────────────────────────────────────────────────
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use soroban_sdk::{testutils::Address as _, testutils::Ledger, Address, Env, String};
-
-    #[test]
-    fn test_initialize() {
-        let env = Env::default();
-        let id = env.register(MarketPayContract, ());
-        let client = MarketPayContractClient::new(&env, &id);
-        let admin = Address::generate(&env);
-        let treasury = Address::generate(&env);
-        client.initialize(&admin, &treasury);
-        assert_eq!(client.get_admin(), admin);
-    }
-
-    #[test]
-    #[should_panic(expected = "Already initialized")]
-    fn test_double_init_panics() {
-        let env = Env::default();
-        let id = env.register(MarketPayContract, ());
-        let c = MarketPayContractClient::new(&env, &id);
-        let admin = Address::generate(&env);
-        let treasury = Address::generate(&env);
-        c.initialize(&admin, &treasury);
-        c.initialize(&admin, &treasury);
-    }
-
-    #[test]
-    fn test_escrow_count_starts_zero() {
-        let env = Env::default();
-        let id = env.register(MarketPayContract, ());
-        let c = MarketPayContractClient::new(&env, &id);
-        let admin = Address::generate(&env);
-        let treasury = Address::generate(&env);
-        c.initialize(&admin, &treasury);
-        assert_eq!(c.get_escrow_count(), 0);
-    }
-
-    #[test]
-    fn test_governance_flow() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let id = env.register(MarketPayContract, ());
-        let client = MarketPayContractClient::new(&env, &id);
-
-        let admin = Address::generate(&env);
-        let treasury = Address::generate(&env);
-        client.initialize(&admin, &treasury);
-
-        let proposer = Address::generate(&env);
-        let voter1 = Address::generate(&env);
-        let voter2 = Address::generate(&env);
-
-        // Give voters completed jobs directly into storage
-        env.as_contract(&id, || {
-            env.storage()
-                .instance()
-                .set(&DataKey::CompletedJobs(voter1.clone()), &1u32);
-            env.storage()
-                .instance()
-                .set(&DataKey::CompletedJobs(voter2.clone()), &1u32);
-        });
-
-        let title = String::from_str(&env, "Test Proposal");
-        let desc = String::from_str(&env, "Description");
-        let pid = client.create_proposal(&proposer, &title, &desc, &100);
-
-        assert_eq!(pid, 1);
-        let prop = client.get_proposal(&pid);
-        assert_eq!(prop.title, title);
-
-        // Vote
-        client.cast_vote(&voter1, &pid, &true);
-        client.cast_vote(&voter2, &pid, &false);
-
-        // Advance ledger using internal testutils sequence setter if possible,
-        // or by generating mock block.
-        // We will mock sequence directly on test env.
-        let mut ledger_info = env.ledger().get();
-        ledger_info.sequence_number += 101;
-        env.ledger().set(ledger_info);
-
-        client.resolve_proposal(&pid);
-
-        let final_prop = client.get_proposal(&pid);
-        assert!(final_prop.resolved);
-        assert!(!final_prop.result); // 1 to 1 is not majority
-    }
-
-    #[test]
-    #[should_panic(expected = "Only users with completed jobs can vote")]
-    fn test_governance_unauthorized_voter() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let id = env.register(MarketPayContract, ());
-        let client = MarketPayContractClient::new(&env, &id);
-
-        let admin = Address::generate(&env);
-        let treasury = Address::generate(&env);
-        client.initialize(&admin, &treasury);
-
-        let proposer = Address::generate(&env);
-        let voter = Address::generate(&env);
-
-        let title = String::from_str(&env, "Test");
-        let desc = String::from_str(&env, "Desc");
-        let pid = client.create_proposal(&proposer, &title, &desc, &100);
-
-        // Panics here
-        client.cast_vote(&voter, &pid, &true);
+        certificates::get_evidence_cids(env, job_id)
     }
 }
 
 #[cfg(test)]
-mod timeout_tests {
-    use super::*;
-    use soroban_sdk::{testutils::Address as _, testutils::Ledger, Address, Env, String};
-
-    fn setup_contract(
-        env: &Env,
-    ) -> (
-        MarketPayContractClient,
-        Address,
-        Address,
-        Address,
-        Address,
-        Address,
-    ) {
-        let id = env.register(MarketPayContract, ());
-        let client = MarketPayContractClient::new(env, &id);
-        let admin = Address::generate(env);
-        let treasury = Address::generate(env);
-        client.initialize(&admin, &treasury);
-
-        let contract_client_addr = Address::generate(env);
-        let freelancer = Address::generate(env);
-        let token_contract = env.register_stellar_asset_contract_v2(admin.clone());
-        let token_id = token_contract.address();
-        let token_admin = token::StellarAssetClient::new(env, &token_id);
-        token_admin.mint(&contract_client_addr, &1000);
-
-        (
-            client,
-            contract_client_addr,
-            freelancer,
-            token_id,
-            admin,
-            treasury,
-        )
-    }
-
-    #[test]
-    fn test_timeout_refund_success() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let (client, contract_client, freelancer, token_id, _admin, _treasury) =
-            setup_contract(&env);
-
-        let job_id = String::from_str(&env, "timeout_job_1");
-        let timeout_ledgers = 10u32;
-        client.create_escrow(
-            &job_id,
-            &contract_client,
-            &CreateEscrowParams {
-                freelancer: freelancer.clone(),
-                token: token_id.clone(),
-                amount: 1000,
-                milestones: None,
-                timeout_ledgers: Some(timeout_ledgers),
-                referrer: None,
-            },
-        );
-
-        let escrow = client.get_escrow(&job_id);
-        assert_eq!(escrow.status, EscrowStatus::Locked);
-        assert_eq!(
-            escrow.timeout_ledger,
-            env.ledger().sequence() + timeout_ledgers
-        );
-
-        // Advance ledger past timeout (both sequence and timestamp)
-        let mut ledger_info = env.ledger().get();
-        ledger_info.sequence_number += timeout_ledgers + 1;
-        ledger_info.timestamp += (DEFAULT_TIMEOUT_SECONDS + 1) as u64; // Advance timestamp too
-        env.ledger().set(ledger_info);
-
-        client.timeout_refund(&job_id, &contract_client);
-
-        let escrow_after = client.get_escrow(&job_id);
-        assert_eq!(escrow_after.status, EscrowStatus::Refunded);
-
-        let token_client = token::Client::new(&env, &token_id);
-        assert_eq!(token_client.balance(&contract_client), 1000);
-    }
-
-    #[test]
-    #[should_panic(expected = "Timeout period has not expired yet")]
-    fn test_timeout_refund_before_timeout_panics() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let (client, contract_client, freelancer, token_id, _admin, _treasury) =
-            setup_contract(&env);
-
-        let job_id = String::from_str(&env, "timeout_job_2");
-        let timeout_ledgers = 100u32;
-        client.create_escrow(
-            &job_id,
-            &contract_client,
-            &CreateEscrowParams {
-                freelancer: freelancer.clone(),
-                token: token_id.clone(),
-                amount: 1000,
-                milestones: None,
-                timeout_ledgers: Some(timeout_ledgers),
-                referrer: None,
-            },
-        );
-
-        // Try to timeout refund before timeout — should panic
-        client.timeout_refund(&job_id, &contract_client);
-    }
-
-    #[test]
-    #[should_panic(expected = "Only the client can request a timeout refund")]
-    fn test_timeout_refund_unauthorized_panics() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let (client, contract_client, freelancer, token_id, _admin, _treasury) =
-            setup_contract(&env);
-
-        let job_id = String::from_str(&env, "timeout_job_3");
-        let timeout_ledgers = 5u32;
-        client.create_escrow(
-            &job_id,
-            &contract_client,
-            &CreateEscrowParams {
-                freelancer: freelancer.clone(),
-                token: token_id.clone(),
-                amount: 1000,
-                milestones: None,
-                timeout_ledgers: Some(timeout_ledgers),
-                referrer: None,
-            },
-        );
-
-        let mut ledger_info = env.ledger().get();
-        ledger_info.sequence_number += timeout_ledgers + 1;
-        env.ledger().set(ledger_info);
-
-        let attacker = Address::generate(&env);
-        client.timeout_refund(&job_id, &attacker);
-    }
-
-    #[test]
-    #[should_panic(expected = "Escrow is not in Locked state")]
-    fn test_timeout_refund_after_start_work_panics() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let (client, contract_client, freelancer, token_id, _admin, _treasury) =
-            setup_contract(&env);
-
-        let job_id = String::from_str(&env, "timeout_job_4");
-        let timeout_ledgers = 10u32;
-        client.create_escrow(
-            &job_id,
-            &contract_client,
-            &CreateEscrowParams {
-                freelancer: freelancer.clone(),
-                token: token_id.clone(),
-                amount: 1000,
-                milestones: None,
-                timeout_ledgers: Some(timeout_ledgers),
-                referrer: None,
-            },
-        );
-
-        // Start work changes status to InProgress (freelancer starts work)
-        client.start_work(&job_id, &freelancer);
-
-        let mut ledger_info = env.ledger().get();
-        ledger_info.sequence_number += timeout_ledgers + 1;
-        env.ledger().set(ledger_info);
-
-        client.timeout_refund(&job_id, &contract_client);
-    }
-
-    #[test]
-    fn test_timeout_refund_with_custom_timeout() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let (client, contract_client, freelancer, token_id, _admin, _treasury) =
-            setup_contract(&env);
-
-        let job_id = String::from_str(&env, "custom_timeout_job");
-        let custom_timeout = 50u32;
-        client.create_escrow(
-            &job_id,
-            &contract_client,
-            &CreateEscrowParams {
-                freelancer: freelancer.clone(),
-                token: token_id.clone(),
-                amount: 500,
-                milestones: None,
-                timeout_ledgers: Some(custom_timeout),
-                referrer: None,
-            },
-        );
-
-        let escrow = client.get_escrow(&job_id);
-        assert_eq!(
-            escrow.timeout_ledger,
-            env.ledger().sequence() + custom_timeout
-        );
-    }
-
-    #[test]
-    fn test_default_timeout_ledgers() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let (client, contract_client, freelancer, token_id, _admin, _treasury) =
-            setup_contract(&env);
-
-        let job_id = String::from_str(&env, "default_timeout_job");
-        client.create_escrow(
-            &job_id,
-            &contract_client,
-            &CreateEscrowParams {
-                freelancer: freelancer.clone(),
-                token: token_id.clone(),
-                amount: 500,
-                milestones: None,
-                timeout_ledgers: None,
-                referrer: None,
-            },
-        );
-
-        let escrow = client.get_escrow(&job_id);
-        assert_eq!(
-            escrow.timeout_ledger,
-            env.ledger().sequence() + DEFAULT_TIMEOUT_LEDGERS
-        );
-    }
-
-    #[test]
-    fn test_get_timeout_ledger() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let (client, contract_client, freelancer, token_id, _admin, _treasury) =
-            setup_contract(&env);
-
-        let job_id = String::from_str(&env, "get_timeout_job");
-        let timeout = 25u32;
-        client.create_escrow(
-            &job_id,
-            &contract_client,
-            &CreateEscrowParams {
-                freelancer: freelancer.clone(),
-                token: token_id.clone(),
-                amount: 500,
-                milestones: None,
-                timeout_ledgers: Some(timeout),
-                referrer: None,
-            },
-        );
-
-        assert_eq!(
-            client.get_timeout_ledger(&job_id),
-            env.ledger().sequence() + timeout
-        );
-    }
-
-    #[test]
-    fn test_timeout_refund_legacy_exact_ledger_success() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let (client, contract_client, freelancer, token_id, _admin, _treasury) =
-            setup_contract(&env);
-
-        let job_id = String::from_str(&env, "legacy_timeout_exact");
-        let timeout_ledgers = 10u32;
-        client.create_escrow(
-            &job_id,
-            &contract_client,
-            &CreateEscrowParams {
-                freelancer: freelancer.clone(),
-                token: token_id.clone(),
-                amount: 1000,
-                milestones: None,
-                timeout_ledgers: Some(timeout_ledgers),
-                referrer: None,
-            },
-        );
-
-        // Remove TimeoutTimestamp to trigger legacy sequence-based fallback
-        env.as_contract(&client.address, || {
-            env.storage()
-                .instance()
-                .remove(&DataKey::TimeoutTimestamp(job_id.clone()));
-        });
-
-        let mut ledger_info = env.ledger().get();
-        ledger_info.sequence_number += timeout_ledgers; // EXACTLY at timeout_ledger
-        env.ledger().set(ledger_info);
-
-        client.timeout_refund(&job_id, &contract_client);
-
-        let escrow_after = client.get_escrow(&job_id);
-        assert_eq!(escrow_after.status, EscrowStatus::Refunded);
-
-        let token_client = token::Client::new(&env, &token_id);
-        assert_eq!(token_client.balance(&contract_client), 1000);
-    }
-
-    #[test]
-    #[should_panic(expected = "Timeout period has not expired yet")]
-    fn test_timeout_refund_legacy_one_ledger_before_failure() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let (client, contract_client, freelancer, token_id, _admin, _treasury) =
-            setup_contract(&env);
-
-        let job_id = String::from_str(&env, "legacy_timeout_before");
-        let timeout_ledgers = 10u32;
-        client.create_escrow(
-            &job_id,
-            &contract_client,
-            &CreateEscrowParams {
-                freelancer: freelancer.clone(),
-                token: token_id.clone(),
-                amount: 1000,
-                milestones: None,
-                timeout_ledgers: Some(timeout_ledgers),
-                referrer: None,
-            },
-        );
-
-        // Remove TimeoutTimestamp to trigger legacy sequence-based fallback
-        env.as_contract(&client.address, || {
-            env.storage()
-                .instance()
-                .remove(&DataKey::TimeoutTimestamp(job_id.clone()));
-        });
-
-        let mut ledger_info = env.ledger().get();
-        ledger_info.sequence_number += timeout_ledgers - 1; // ONE ledger before timeout_ledger
-        env.ledger().set(ledger_info);
-
-        client.timeout_refund(&job_id, &contract_client);
-    }
-
-    #[test]
-    #[should_panic(expected = "Escrow is not in Locked state")]
-    fn test_concurrent_release_and_timeout_refund_release_first() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let (client, contract_client, freelancer, token_id, _admin, _treasury) =
-            setup_contract(&env);
-
-        let job_id = String::from_str(&env, "concurrent_release_first");
-        let timeout_ledgers = 10u32;
-        client.create_escrow(
-            &job_id,
-            &contract_client,
-            &CreateEscrowParams {
-                freelancer: freelancer.clone(),
-                token: token_id.clone(),
-                amount: 1000,
-                milestones: None,
-                timeout_ledgers: Some(timeout_ledgers),
-                referrer: None,
-            },
-        );
-
-        // Advance ledger past timeout (both sequence and timestamp)
-        let mut ledger_info = env.ledger().get();
-        ledger_info.sequence_number += timeout_ledgers + 1;
-        ledger_info.timestamp += (DEFAULT_TIMEOUT_SECONDS + 1) as u64;
-        env.ledger().set(ledger_info);
-
-        // First action: Release Escrow (succeeds)
-        client.release_escrow(&job_id, &contract_client);
-
-        let escrow_after = client.get_escrow(&job_id);
-        assert_eq!(escrow_after.status, EscrowStatus::Released);
-
-        // Second action: Timeout Refund (fails because status is no longer Locked)
-        client.timeout_refund(&job_id, &contract_client);
-    }
-
-    #[test]
-    #[should_panic(expected = "Cannot release escrow in current status")]
-    fn test_concurrent_release_and_timeout_refund_timeout_first() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let (client, contract_client, freelancer, token_id, _admin, _treasury) =
-            setup_contract(&env);
-
-        let job_id = String::from_str(&env, "concurrent_timeout_first");
-        let timeout_ledgers = 10u32;
-        client.create_escrow(
-            &job_id,
-            &contract_client,
-            &CreateEscrowParams {
-                freelancer: freelancer.clone(),
-                token: token_id.clone(),
-                amount: 1000,
-                milestones: None,
-                timeout_ledgers: Some(timeout_ledgers),
-                referrer: None,
-            },
-        );
-
-        // Advance ledger past timeout (both sequence and timestamp)
-        let mut ledger_info = env.ledger().get();
-        ledger_info.sequence_number += timeout_ledgers + 1;
-        ledger_info.timestamp += (DEFAULT_TIMEOUT_SECONDS + 1) as u64;
-        env.ledger().set(ledger_info);
-
-        // First action: Timeout Refund (succeeds)
-        client.timeout_refund(&job_id, &contract_client);
-
-        let escrow_after = client.get_escrow(&job_id);
-        assert_eq!(escrow_after.status, EscrowStatus::Refunded);
-
-        // Second action: Release Escrow (fails because status is no longer InProgress/Locked)
-        client.release_escrow(&job_id, &contract_client);
-    }
-}
-
-#[cfg(test)]
-mod regression_tests {
-    use super::*;
-    use soroban_sdk::{testutils::Address as _, Address, Env, String, Vec};
-
-    #[test]
-    fn test_release_escrow_state_consistency_regression() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let id = env.register(MarketPayContract, ());
-        let contract_client = MarketPayContractClient::new(&env, &id);
-
-        let admin = Address::generate(&env);
-        let treasury = Address::generate(&env);
-        contract_client.initialize(&admin, &treasury);
-
-        let client = Address::generate(&env);
-        let freelancer = Address::generate(&env);
-
-        let token_contract = env.register_stellar_asset_contract_v2(admin.clone());
-        let token_id = token_contract.address();
-        let token_client = token::Client::new(&env, &token_id);
-        let token_admin = token::StellarAssetClient::new(&env, &token_id);
-        token_admin.mint(&client, &1000);
-
-        let job_id = String::from_str(&env, "job1");
-        contract_client.create_escrow(
-            &job_id,
-            &client.clone(),
-            &CreateEscrowParams {
-                freelancer: freelancer.clone(),
-                token: token_id.clone(),
-                amount: 1000,
-                milestones: None,
-                timeout_ledgers: None,
-                referrer: None,
-            },
-        );
-        contract_client.start_work(&job_id, &freelancer.clone());
-
-        contract_client.release_escrow(&job_id, &client.clone());
-
-        let escrow = contract_client.get_escrow(&job_id);
-        assert_eq!(escrow.status, EscrowStatus::Released);
-        // Fee = 1000 * 1% = 10. Freelancer gets 990.
-        assert_eq!(token_client.balance(&freelancer), 990);
-    }
-
-    #[test]
-    fn test_release_with_conversion() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let id = env.register(MarketPayContract, ());
-        let contract_client = MarketPayContractClient::new(&env, &id);
-
-        let admin = Address::generate(&env);
-        let treasury = Address::generate(&env);
-        contract_client.initialize(&admin, &treasury);
-
-        let client = Address::generate(&env);
-        let freelancer = Address::generate(&env);
-
-        let token_contract = env.register_stellar_asset_contract_v2(admin.clone());
-        let token_id = token_contract.address();
-        let token_admin = token::StellarAssetClient::new(&env, &token_id);
-        token_admin.mint(&client, &1000);
-
-        let job_id = String::from_str(&env, "job_conv");
-        contract_client.create_escrow(
-            &job_id,
-            &client.clone(),
-            &CreateEscrowParams {
-                freelancer: freelancer.clone(),
-                token: token_id.clone(),
-                amount: 1000,
-                milestones: None,
-                timeout_ledgers: None,
-                referrer: None,
-            },
-        );
-
-        let target_token = Address::generate(&env);
-        contract_client.release_with_conversion(&job_id, &client.clone(), &target_token, &900);
-
-        let escrow = contract_client.get_escrow(&job_id);
-        assert_eq!(escrow.status, EscrowStatus::Released);
-    }
-
-    #[test]
-    fn test_partial_release() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let id = env.register(MarketPayContract, ());
-        let contract_client = MarketPayContractClient::new(&env, &id);
-
-        let admin = Address::generate(&env);
-        let treasury = Address::generate(&env);
-        contract_client.initialize(&admin, &treasury);
-
-        let client = Address::generate(&env);
-        let freelancer = Address::generate(&env);
-
-        let token_contract = env.register_stellar_asset_contract_v2(admin.clone());
-        let token_id = token_contract.address();
-        let token_client = token::Client::new(&env, &token_id);
-        let token_admin = token::StellarAssetClient::new(&env, &token_id);
-        token_admin.mint(&client, &1000);
-
-        let mut milestones = Vec::new(&env);
-        milestones.push_back(MilestoneInput {
-            description: String::from_str(&env, "Design"),
-            percentage: 40,
-        });
-        milestones.push_back(MilestoneInput {
-            description: String::from_str(&env, "Build"),
-            percentage: 60,
-        });
-
-        let job_id = String::from_str(&env, "job_partial");
-        contract_client.create_escrow(
-            &job_id,
-            &client.clone(),
-            &CreateEscrowParams {
-                freelancer: freelancer.clone(),
-                token: token_id.clone(),
-                amount: 1000,
-                milestones: Some(milestones),
-                timeout_ledgers: None,
-                referrer: None,
-            },
-        );
-        contract_client.start_work(&job_id, &freelancer.clone());
-
-        // Raise dispute to test that we can still release milestones
-        contract_client.raise_dispute(&job_id, &client.clone());
-
-        contract_client.release_milestone(&job_id, &0u32, &client.clone());
-
-        let escrow = contract_client.get_escrow(&job_id);
-        assert_eq!(escrow.status, EscrowStatus::Disputed);
-        // Milestone 0 = 40% = 400. Fee = 400 * 1% = 4. Freelancer gets 396.
-        assert_eq!(token_client.balance(&freelancer), 396);
-        assert!(escrow.milestones.get(0).unwrap().released);
-        assert!(!escrow.milestones.get(1).unwrap().released);
-
-        // Release final milestone
-        contract_client.release_milestone(&job_id, &1u32, &client.clone());
-        let escrow2 = contract_client.get_escrow(&job_id);
-        assert_eq!(escrow2.status, EscrowStatus::Released);
-        // Remaining 60% = 600. Fee = 600 * 1% = 6. Freelancer gets 594.
-        // Total = 396 + 594 = 990
-        assert_eq!(token_client.balance(&freelancer), 990);
-    }
-}
-
-#[cfg(test)]
-mod upgrade_tests {
-    use super::*;
-    use soroban_sdk::{testutils::Address as _, Address, BytesN, Env, String};
-
-    /// Verifies that:
-    ///   - get_version() returns 1 after initialize()
-    ///   - upgrade() with a valid hash increments the version to 2
-    ///   - existing escrow state is preserved after upgrade
-    ///
-    /// Note: `update_current_contract_wasm` requires the hash to reference
-    /// an installed WASM blob. In unit tests we verify the auth guard and
-    /// version-bump logic; the actual WASM swap is covered by integration /
-    /// testnet tests (see README upgrade process).
-    #[test]
-    fn test_version_starts_at_one() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let id = env.register(MarketPayContract, ());
-        let client = MarketPayContractClient::new(&env, &id);
-        let admin = Address::generate(&env);
-        let treasury = Address::generate(&env);
-        client.initialize(&admin, &treasury);
-        assert_eq!(client.get_version(), 1u32);
-    }
-
-    /// Verifies escrow state is readable before and after a simulated upgrade
-    /// (version bump via direct storage write, bypassing WASM swap).
-    #[test]
-    fn test_escrow_state_preserved_across_version_bump() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let id = env.register(MarketPayContract, ());
-        let client = MarketPayContractClient::new(&env, &id);
-
-        let admin = Address::generate(&env);
-        let treasury = Address::generate(&env);
-        client.initialize(&admin, &treasury);
-
-        let depositor = Address::generate(&env);
-        let freelancer = Address::generate(&env);
-        let token_contract = env.register_stellar_asset_contract_v2(admin.clone());
-        let token_id = token_contract.address();
-        let token_admin = token::StellarAssetClient::new(&env, &token_id);
-        token_admin.mint(&depositor, &500);
-
-        let job_id = String::from_str(&env, "upgrade_job_1");
-        client.create_escrow(
-            &job_id,
-            &depositor,
-            &CreateEscrowParams {
-                freelancer: freelancer.clone(),
-                token: token_id.clone(),
-                amount: 500,
-                milestones: None,
-                timeout_ledgers: None,
-                referrer: None,
-            },
-        );
-
-        // Simulate the version bump that upgrade() performs (without WASM swap)
-        env.as_contract(&id, || {
-            let v: u32 = env.storage().instance().get(&DataKey::Version).unwrap_or(1);
-            env.storage().instance().set(&DataKey::Version, &(v + 1));
-        });
-
-        assert_eq!(client.get_version(), 2u32);
-
-        // Escrow state intact
-        let escrow = client.get_escrow(&job_id);
-        assert_eq!(escrow.amount, 500);
-        assert_eq!(escrow.status, EscrowStatus::Locked);
-    }
-
-    #[test]
-    #[should_panic]
-    fn test_upgrade_rejected_for_non_admin() {
-        let env = Env::default();
-        // Do NOT mock_all_auths — auth will fail for non-admin
-        let id = env.register(MarketPayContract, ());
-        let client = MarketPayContractClient::new(&env, &id);
-
-        let admin = Address::generate(&env);
-        let treasury = Address::generate(&env);
-        client.initialize(&admin, &treasury);
-
-        let fake_hash = BytesN::from_array(&env, &[0u8; 32]);
-        // Called without admin auth → should panic
-        client.upgrade(&fake_hash);
-    }
-
-    #[test]
-    fn test_get_milestone_returns_correct_milestone() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let id = env.register(MarketPayContract, ());
-        let client = MarketPayContractClient::new(&env, &id);
-        let admin = Address::generate(&env);
-        client.initialize(&admin, &admin);
-
-        let depositor = Address::generate(&env);
-        let freelancer = Address::generate(&env);
-        let token_contract = env.register_stellar_asset_contract_v2(admin.clone());
-        let token_id = token_contract.address();
-        let token_admin = token::StellarAssetClient::new(&env, &token_id);
-        token_admin.mint(&depositor, &1000);
-
-        let mut ms = Vec::new(&env);
-        ms.push_back(MilestoneInput {
-            description: String::from_str(&env, "Phase 1"),
-            percentage: 40,
-        });
-        ms.push_back(MilestoneInput {
-            description: String::from_str(&env, "Phase 2"),
-            percentage: 60,
-        });
-
-        let job_id = String::from_str(&env, "ms-getter-1");
-        client.create_escrow(
-            &job_id,
-            &depositor,
-            &CreateEscrowParams {
-                freelancer: freelancer.clone(),
-                token: token_id.clone(),
-                amount: 1000,
-                milestones: Some(ms),
-                timeout_ledgers: None,
-                referrer: None,
-            },
-        );
-
-        let ms0 = client.get_milestone(&job_id, &0u32);
-        assert_eq!(ms0.id, 0u32);
-        assert_eq!(ms0.percentage, 40u32);
-
-        let ms1 = client.get_milestone(&job_id, &1u32);
-        assert_eq!(ms1.id, 1u32);
-        assert_eq!(ms1.percentage, 60u32);
-    }
-
-    #[test]
-    #[should_panic(expected = "Milestone index out of bounds")]
-    fn test_get_milestone_out_of_bounds_panics() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let id = env.register(MarketPayContract, ());
-        let client = MarketPayContractClient::new(&env, &id);
-        let admin = Address::generate(&env);
-        client.initialize(&admin, &admin);
-
-        let depositor = Address::generate(&env);
-        let freelancer = Address::generate(&env);
-        let token_contract = env.register_stellar_asset_contract_v2(admin.clone());
-        let token_id = token_contract.address();
-        let token_admin = token::StellarAssetClient::new(&env, &token_id);
-        token_admin.mint(&depositor, &500);
-
-        let mut ms = Vec::new(&env);
-        ms.push_back(MilestoneInput {
-            description: String::from_str(&env, "Only milestone"),
-            percentage: 100,
-        });
-
-        let job_id = String::from_str(&env, "ms-oob-1");
-        client.create_escrow(
-            &job_id,
-            &depositor,
-            &CreateEscrowParams {
-                freelancer: freelancer.clone(),
-                token: token_id.clone(),
-                amount: 500,
-                milestones: Some(ms),
-                timeout_ledgers: None,
-                referrer: None,
-            },
-        );
-
-        client.get_milestone(&job_id, &5u32);
-    }
-
-    #[test]
-    fn test_is_frozen_defaults_false() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let id = env.register(MarketPayContract, ());
-        let client = MarketPayContractClient::new(&env, &id);
-        let admin = Address::generate(&env);
-        client.initialize(&admin, &admin);
-        assert!(!client.is_frozen());
-    }
-}
-
-#[cfg(test)]
-mod sealed_bid_tests {
-    use super::*;
-    use soroban_sdk::{
-        testutils::Address as _, testutils::Ledger, Address, Bytes, BytesN, Env, String,
-    };
-
-    fn bid_commitment(env: &Env, amount: i128, nonce: BytesN<32>) -> BytesN<32> {
-        let mut payload = Bytes::new(env);
-        for byte in amount.to_be_bytes().iter() {
-            payload.push_back(*byte);
-        }
-        for byte in nonce.to_array().iter() {
-            payload.push_back(*byte);
-        }
-        env.crypto().sha256(&payload).into()
-    }
-
-    fn setup(env: &Env) -> (Address, MarketPayContractClient, Address, Address, String) {
-        env.mock_all_auths();
-        let id = env.register(MarketPayContract, ());
-        let client = MarketPayContractClient::new(env, &id);
-        let admin = Address::generate(env);
-        let treasury = Address::generate(env);
-        let owner = Address::generate(env);
-        client.initialize(&admin, &treasury);
-        let job_id = String::from_str(env, "sealed-bid-job-1");
-        client.commit_budget(&job_id, &1_000, &owner);
-        (id, client, owner, admin, job_id)
-    }
-
-    #[test]
-    fn test_reveal_bid_verifies_commitment() {
-        let env = Env::default();
-
-        let (_id, client, owner, _admin, job_id) = setup(&env);
-        let freelancer = Address::generate(&env);
-        let nonce = BytesN::from_array(&env, &[7u8; 32]);
-        let amount = 450i128;
-        let commitment = bid_commitment(&env, amount, nonce.clone());
-
-        client.submit_bid_commitment(&job_id, &freelancer, &commitment);
-        client.close_bidding(&job_id, &owner);
-        client.reveal_bid(&job_id, &freelancer, &amount, &nonce);
-
-        let reveals = client.get_revealed_bids(&job_id);
-        assert_eq!(reveals.len(), 1);
-        let revealed = reveals.get(0).unwrap();
-        assert_eq!(revealed.amount, amount);
-        assert_eq!(revealed.freelancer, freelancer);
-    }
-
-    #[test]
-    #[should_panic(expected = "Commitment verification failed")]
-    fn test_reveal_bid_with_invalid_nonce_rejected() {
-        let env = Env::default();
-        let (_id, client, owner, _admin, job_id) = setup(&env);
-        let freelancer = Address::generate(&env);
-        let amount = 500i128;
-        let nonce = BytesN::from_array(&env, &[1u8; 32]);
-        let bad_nonce = BytesN::from_array(&env, &[2u8; 32]);
-        let commitment = bid_commitment(&env, amount, nonce);
-
-        client.submit_bid_commitment(&job_id, &freelancer, &commitment);
-        client.close_bidding(&job_id, &owner);
-        client.reveal_bid(&job_id, &freelancer, &amount, &bad_nonce);
-    }
-
-    #[test]
-    #[should_panic(expected = "Reveal window has closed")]
-    fn test_late_reveal_rejected() {
-        let env = Env::default();
-        let (id, client, owner, _admin, job_id) = setup(&env);
-        let freelancer = Address::generate(&env);
-        let nonce = BytesN::from_array(&env, &[3u8; 32]);
-        let amount = 525i128;
-        let commitment = bid_commitment(&env, amount, nonce.clone());
-
-        client.submit_bid_commitment(&job_id, &freelancer, &commitment);
-        client.close_bidding(&job_id, &owner);
-
-        // Extend instance storage TTL so it survives the ledger jump below.
-        env.as_contract(&id, || {
-            env.storage().instance().extend_ttl(20_000, 20_000);
-        });
-
-        let mut ledger = env.ledger().get();
-        ledger.sequence_number += REVEAL_WINDOW_LEDGERS + 1;
-        env.ledger().set(ledger);
-
-        client.reveal_bid(&job_id, &freelancer, &amount, &nonce);
-    }
-}
-
-#[cfg(test)]
-mod deliverable_oracle_tests {
-    use super::*;
-    use soroban_sdk::{testutils::Address as _, Address, BytesN, Env, String};
-
-    fn setup(env: &Env) -> (MarketPayContractClient, Address, Address, Address, Address) {
-        env.mock_all_auths();
-        let id = env.register(MarketPayContract, ());
-        let contract = MarketPayContractClient::new(env, &id);
-        let admin = Address::generate(env);
-        let treasury = Address::generate(env);
-        contract.initialize(&admin, &treasury);
-
-        let client = Address::generate(env);
-        let freelancer = Address::generate(env);
-        let token_contract = env.register_stellar_asset_contract_v2(admin.clone());
-        let token_id = token_contract.address();
-        let token_admin = token::StellarAssetClient::new(env, &token_id);
-        token_admin.mint(&client, &1_000);
-
-        (contract, admin, client, freelancer, token_id)
-    }
-
-    #[test]
-    fn test_submit_deliverable_match_auto_releases() {
-        let env = Env::default();
-        let (contract, _admin, client, freelancer, token_id) = setup(&env);
-        let job_id = String::from_str(&env, "deliverable-match");
-        let expected_hash = BytesN::from_array(&env, &[9u8; 32]);
-
-        contract.create_escrow_with_deliverable(
-            &job_id,
-            &client,
-            &CreateEscrowParams {
-                freelancer: freelancer.clone(),
-                token: token_id.clone(),
-                amount: 1_000,
-                milestones: None,
-                timeout_ledgers: None,
-                referrer: None,
-            },
-            &expected_hash,
-        );
-
-        contract.submit_deliverable(&job_id, &expected_hash, &freelancer);
-
-        let escrow = contract.get_escrow(&job_id);
-        assert_eq!(escrow.status, EscrowStatus::Released);
-
-        let token_client = token::Client::new(&env, &token_id);
-        // Fee = 1000 * 1% = 10. Freelancer gets 990.
-        assert_eq!(token_client.balance(&freelancer), 990);
-    }
-
-    #[test]
-    fn test_submit_deliverable_mismatch_enters_dispute() {
-        let env = Env::default();
-        let (contract, _admin, client, freelancer, token_id) = setup(&env);
-        let job_id = String::from_str(&env, "deliverable-mismatch");
-        let expected_hash = BytesN::from_array(&env, &[1u8; 32]);
-        let actual_hash = BytesN::from_array(&env, &[2u8; 32]);
-
-        contract.create_escrow_with_deliverable(
-            &job_id,
-            &client,
-            &CreateEscrowParams {
-                freelancer: freelancer.clone(),
-                token: token_id.clone(),
-                amount: 1_000,
-                milestones: None,
-                timeout_ledgers: None,
-                referrer: None,
-            },
-            &expected_hash,
-        );
-
-        contract.submit_deliverable(&job_id, &actual_hash, &freelancer);
-        let escrow = contract.get_escrow(&job_id);
-        assert_eq!(escrow.status, EscrowStatus::Disputed);
-
-        let token_client = token::Client::new(&env, &token_id);
-        assert_eq!(token_client.balance(&freelancer), 0);
-    }
-}
-
-#[cfg(test)]
-mod milestone_pct_tests {
-    use super::*;
-    use soroban_sdk::{testutils::Address as _, Address, Env, String, Vec};
-
-    fn setup(env: &Env) -> (MarketPayContractClient, Address, Address, Address) {
-        env.mock_all_auths();
-        let id = env.register(MarketPayContract, ());
-        let contract = MarketPayContractClient::new(env, &id);
-        let admin = Address::generate(env);
-        let treasury = Address::generate(env);
-        contract.initialize(&admin, &treasury);
-
-        let client = Address::generate(env);
-        let freelancer = Address::generate(env);
-        let token_contract = env.register_stellar_asset_contract_v2(admin.clone());
-        let token_id = token_contract.address();
-        let token_admin = token::StellarAssetClient::new(env, &token_id);
-        token_admin.mint(&client, &1_000);
-
-        (contract, client, freelancer, token_id)
-    }
-
-    #[test]
-    fn test_create_escrow_with_milestones_valid() {
-        let env = Env::default();
-        let (contract, client, freelancer, token_id) = setup(&env);
-        let job_id = String::from_str(&env, "ms-job-1");
-
-        let mut ms = Vec::new(&env);
-        ms.push_back(MilestoneInput {
-            description: String::from_str(&env, "Phase 1"),
-            percentage: 40,
-        });
-        ms.push_back(MilestoneInput {
-            description: String::from_str(&env, "Phase 2"),
-            percentage: 60,
-        });
-
-        contract.create_escrow_with_milestones(
-            &job_id,
-            &client,
-            &CreateEscrowParams {
-                freelancer: freelancer.clone(),
-                token: token_id.clone(),
-                amount: 1_000,
-                milestones: Some(ms),
-                timeout_ledgers: None,
-                referrer: None,
-            },
-        );
-
-        let escrow = contract.get_escrow(&job_id);
-        assert_eq!(escrow.milestones.len(), 2);
-    }
-
-    #[test]
-    #[should_panic(expected = "Milestone percentages must sum to 100")]
-    fn test_invalid_percentages_rejected() {
-        let env = Env::default();
-        let (contract, client, freelancer, token_id) = setup(&env);
-        let job_id = String::from_str(&env, "ms-job-2");
-
-        let mut ms = Vec::new(&env);
-        ms.push_back(MilestoneInput {
-            description: String::from_str(&env, "Phase 1"),
-            percentage: 40,
-        });
-        ms.push_back(MilestoneInput {
-            description: String::from_str(&env, "Phase 2"),
-            percentage: 50,
-        });
-
-        contract.create_escrow_with_milestones(
-            &job_id,
-            &client,
-            &CreateEscrowParams {
-                freelancer: freelancer.clone(),
-                token: token_id.clone(),
-                amount: 1_000,
-                milestones: Some(ms),
-                timeout_ledgers: None,
-                referrer: None,
-            },
-        );
-    }
-
-    #[test]
-    fn test_release_first_milestone() {
-        let env = Env::default();
-        let (contract, client, freelancer, token_id) = setup(&env);
-        let job_id = String::from_str(&env, "ms-job-3");
-
-        let mut ms = Vec::new(&env);
-        ms.push_back(MilestoneInput {
-            description: String::from_str(&env, "Phase 1"),
-            percentage: 40,
-        });
-        ms.push_back(MilestoneInput {
-            description: String::from_str(&env, "Phase 2"),
-            percentage: 60,
-        });
-
-        contract.create_escrow_with_milestones(
-            &job_id,
-            &client,
-            &CreateEscrowParams {
-                freelancer: freelancer.clone(),
-                token: token_id.clone(),
-                amount: 1_000,
-                milestones: Some(ms),
-                timeout_ledgers: None,
-                referrer: None,
-            },
-        );
-        contract.start_work(&job_id, &freelancer);
-        contract.release_milestone(&job_id, &0u32, &client);
-
-        let token_client = token::Client::new(&env, &token_id);
-        // Milestone 0 = 40% = 400. Fee = 400 * 1% = 4. Freelancer gets 396.
-        assert_eq!(token_client.balance(&freelancer), 396);
-
-        let escrow = contract.get_escrow(&job_id);
-        assert_eq!(escrow.status, EscrowStatus::InProgress);
-        assert!(escrow.milestones.get(0).unwrap().released);
-    }
-
-    #[test]
-    fn test_release_all_milestones_marks_released() {
-        let env = Env::default();
-        let (contract, client, freelancer, token_id) = setup(&env);
-        let job_id = String::from_str(&env, "ms-job-4");
-
-        let mut ms = Vec::new(&env);
-        ms.push_back(MilestoneInput {
-            description: String::from_str(&env, "Phase 1"),
-            percentage: 40,
-        });
-        ms.push_back(MilestoneInput {
-            description: String::from_str(&env, "Phase 2"),
-            percentage: 60,
-        });
-
-        contract.create_escrow_with_milestones(
-            &job_id,
-            &client,
-            &CreateEscrowParams {
-                freelancer: freelancer.clone(),
-                token: token_id.clone(),
-                amount: 1_000,
-                milestones: Some(ms),
-                timeout_ledgers: None,
-                referrer: None,
-            },
-        );
-        contract.start_work(&job_id, &freelancer);
-        contract.release_milestone(&job_id, &0u32, &client);
-        contract.release_milestone(&job_id, &1u32, &client);
-
-        let token_client = token::Client::new(&env, &token_id);
-        // Total fee = 1000 * 1% = 10. Freelancer gets 990.
-        assert_eq!(token_client.balance(&freelancer), 990);
-
-        let escrow = contract.get_escrow(&job_id);
-        assert_eq!(escrow.status, EscrowStatus::Released);
-    }
-}
-
-#[cfg(test)]
-mod deliverable_hash_tests {
-    use super::*;
-    use soroban_sdk::{testutils::Address as _, Address, BytesN, Env, String};
-
-    fn setup(env: &Env) -> (MarketPayContractClient, Address, Address, Address) {
-        env.mock_all_auths();
-        let id = env.register(MarketPayContract, ());
-        let contract = MarketPayContractClient::new(env, &id);
-        let admin = Address::generate(env);
-        contract.initialize(&admin, &admin);
-        contract.set_platform_fee_bps(&admin, &0);
-
-        let client = Address::generate(env);
-        let freelancer = Address::generate(env);
-        let token_contract = env.register_stellar_asset_contract_v2(admin.clone());
-        let token_id = token_contract.address();
-        let token_admin = token::StellarAssetClient::new(env, &token_id);
-        token_admin.mint(&client, &1_000);
-
-        (contract, client, freelancer, token_id)
-    }
-
-    #[test]
-    fn test_freelancer_submits_deliverable_hash() {
-        let env = Env::default();
-        let (contract, client, freelancer, token_id) = setup(&env);
-        let job_id = String::from_str(&env, "dh-job-1");
-        let expected_hash = BytesN::from_array(&env, &[0xabu8; 32]);
-
-        contract.create_escrow_with_deliverable(
-            &job_id,
-            &client,
-            &CreateEscrowParams {
-                freelancer: freelancer.clone(),
-                token: token_id.clone(),
-                amount: 1_000,
-                milestones: None,
-                timeout_ledgers: None,
-                referrer: None,
-            },
-            &expected_hash,
-        );
-
-        contract.submit_deliverable_hash(&job_id, &freelancer, &expected_hash);
-
-        let stored = contract.get_freelancer_deliverable_hash(&job_id);
-        assert_eq!(stored, Some(expected_hash));
-    }
-
-    #[test]
-    fn test_release_succeeds_with_matching_hash() {
-        let env = Env::default();
-        let (contract, client, freelancer, token_id) = setup(&env);
-        let job_id = String::from_str(&env, "dh-job-2");
-        let expected_hash = BytesN::from_array(&env, &[0xabu8; 32]);
-
-        contract.create_escrow_with_deliverable(
-            &job_id,
-            &client,
-            &CreateEscrowParams {
-                freelancer: freelancer.clone(),
-                token: token_id.clone(),
-                amount: 1_000,
-                milestones: None,
-                timeout_ledgers: None,
-                referrer: None,
-            },
-            &expected_hash,
-        );
-
-        contract.submit_deliverable_hash(&job_id, &freelancer, &expected_hash);
-        contract.release_escrow(&job_id, &client);
-
-        let escrow = contract.get_escrow(&job_id);
-        assert_eq!(escrow.status, EscrowStatus::Released);
-        let token_client = token::Client::new(&env, &token_id);
-        assert_eq!(token_client.balance(&freelancer), 1_000);
-    }
-
-    #[test]
-    #[should_panic(expected = "Freelancer deliverable hash does not match or not submitted")]
-    fn test_release_panics_without_submitting_hash() {
-        let env = Env::default();
-        let (contract, client, freelancer, token_id) = setup(&env);
-        let job_id = String::from_str(&env, "dh-job-3");
-        let expected_hash = BytesN::from_array(&env, &[0xabu8; 32]);
-
-        contract.create_escrow_with_deliverable(
-            &job_id,
-            &client,
-            &CreateEscrowParams {
-                freelancer: freelancer.clone(),
-                token: token_id.clone(),
-                amount: 1_000,
-                milestones: None,
-                timeout_ledgers: None,
-                referrer: None,
-            },
-            &expected_hash,
-        );
-
-        contract.release_escrow(&job_id, &client);
-    }
-
-    #[test]
-    #[should_panic(expected = "Freelancer deliverable hash does not match or not submitted")]
-    fn test_release_panics_with_wrong_hash() {
-        let env = Env::default();
-        let (contract, client, freelancer, token_id) = setup(&env);
-        let job_id = String::from_str(&env, "dh-job-4");
-        let expected_hash = BytesN::from_array(&env, &[0xabu8; 32]);
-        let wrong_hash = BytesN::from_array(&env, &[0xbbu8; 32]);
-
-        contract.create_escrow_with_deliverable(
-            &job_id,
-            &client,
-            &CreateEscrowParams {
-                freelancer: freelancer.clone(),
-                token: token_id.clone(),
-                amount: 1_000,
-                milestones: None,
-                timeout_ledgers: None,
-                referrer: None,
-            },
-            &expected_hash,
-        );
-
-        contract.submit_deliverable_hash(&job_id, &freelancer, &wrong_hash);
-        contract.release_escrow(&job_id, &client);
-    }
-
-    #[test]
-    fn test_release_without_expected_hash_works() {
-        let env = Env::default();
-        let (contract, client, freelancer, token_id) = setup(&env);
-        let job_id = String::from_str(&env, "dh-job-5");
-
-        contract.create_escrow(
-            &job_id,
-            &client,
-            &CreateEscrowParams {
-                freelancer: freelancer.clone(),
-                token: token_id.clone(),
-                amount: 1_000,
-                milestones: None,
-                timeout_ledgers: None,
-                referrer: None,
-            },
-        );
-        contract.start_work(&job_id, &freelancer);
-        contract.release_escrow(&job_id, &client);
-
-        let escrow = contract.get_escrow(&job_id);
-        assert_eq!(escrow.status, EscrowStatus::Released);
-    }
-
-    #[test]
-    fn test_get_freelancer_hash_returns_none_when_empty() {
-        let env = Env::default();
-        let (contract, client, freelancer, token_id) = setup(&env);
-        let job_id = String::from_str(&env, "dh-job-6");
-
-        contract.create_escrow(
-            &job_id,
-            &client,
-            &CreateEscrowParams {
-                freelancer: freelancer.clone(),
-                token: token_id.clone(),
-                amount: 1_000,
-                milestones: None,
-                timeout_ledgers: None,
-                referrer: None,
-            },
-        );
-
-        let stored = contract.get_freelancer_deliverable_hash(&job_id);
-        assert_eq!(stored, None);
-    }
-
-    #[test]
-    fn test_verify_deliverable_hash_returns_true_when_match() {
-        let env = Env::default();
-        let (contract, client, freelancer, token_id) = setup(&env);
-        let job_id = String::from_str(&env, "dh-job-7");
-        let expected_hash = BytesN::from_array(&env, &[0xabu8; 32]);
-
-        contract.create_escrow_with_deliverable(
-            &job_id,
-            &client,
-            &CreateEscrowParams {
-                freelancer: freelancer.clone(),
-                token: token_id.clone(),
-                amount: 1_000,
-                milestones: None,
-                timeout_ledgers: None,
-                referrer: None,
-            },
-            &expected_hash,
-        );
-
-        contract.submit_deliverable_hash(&job_id, &freelancer, &expected_hash);
-
-        assert!(contract.verify_deliverable_hash(&job_id));
-    }
-
-    #[test]
-    fn test_verify_deliverable_hash_false_when_not_submitted() {
-        let env = Env::default();
-        let (contract, client, freelancer, token_id) = setup(&env);
-        let job_id = String::from_str(&env, "dh-job-8");
-        let expected_hash = BytesN::from_array(&env, &[0xabu8; 32]);
-
-        contract.create_escrow_with_deliverable(
-            &job_id,
-            &client,
-            &CreateEscrowParams {
-                freelancer: freelancer.clone(),
-                token: token_id.clone(),
-                amount: 1_000,
-                milestones: None,
-                timeout_ledgers: None,
-                referrer: None,
-            },
-            &expected_hash,
-        );
-
-        assert!(!contract.verify_deliverable_hash(&job_id));
-    }
-
-    #[test]
-    #[should_panic(expected = "Only the freelancer can submit deliverable hash")]
-    fn test_non_freelancer_cannot_submit_hash() {
-        let env = Env::default();
-        let (contract, client, freelancer, token_id) = setup(&env);
-        let job_id = String::from_str(&env, "dh-job-9");
-        let h = BytesN::from_array(&env, &[0xabu8; 32]);
-
-        contract.create_escrow_with_deliverable(
-            &job_id,
-            &client,
-            &CreateEscrowParams {
-                freelancer: freelancer.clone(),
-                token: token_id.clone(),
-                amount: 1_000,
-                milestones: None,
-                timeout_ledgers: None,
-                referrer: None,
-            },
-            &h,
-        );
-
-        contract.submit_deliverable_hash(&job_id, &client, &h);
-    }
-}
-
-#[cfg(test)]
-mod extension_tests {
-    use super::*;
-    use soroban_sdk::testutils::Address as _;
-    use soroban_sdk::{Address, Env, String};
-
-    fn setup(env: &Env) -> (MarketPayContractClient, Address, Address, Address) {
-        env.mock_all_auths();
-        let id = env.register(MarketPayContract, ());
-        let client = MarketPayContractClient::new(env, &id);
-        let admin = Address::generate(env);
-        client.initialize(&admin, &admin);
-
-        let contract_client = Address::generate(env);
-        let freelancer = Address::generate(env);
-        let token_contract = env.register_stellar_asset_contract_v2(admin.clone());
-        let token_id = token_contract.address();
-        let token_admin = token::StellarAssetClient::new(env, &token_id);
-        token_admin.mint(&contract_client, &1000);
-
-        (client, contract_client, freelancer, token_id)
-    }
-
-    #[test]
-    fn test_client_requests_extension() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let (client, contract_client, freelancer, token_id) = setup(&env);
-        let job_id = String::from_str(&env, "ext-1");
-        let timeout_ledgers = 10u32;
-
-        client.create_escrow(
-            &job_id,
-            &contract_client,
-            &CreateEscrowParams {
-                freelancer: freelancer.clone(),
-                token: token_id.clone(),
-                amount: 500,
-                milestones: None,
-                timeout_ledgers: Some(timeout_ledgers),
-                referrer: None,
-            },
-        );
-
-        let escrow = client.get_escrow(&job_id);
-        let current_ledger = env.ledger().sequence();
-        let new_timeout = current_ledger + timeout_ledgers + 20;
-
-        client.request_extension(&job_id, &contract_client, &new_timeout);
-
-        let req = client.get_extension_request(&job_id).unwrap();
-        assert_eq!(req.requested_by, contract_client);
-        assert_eq!(req.new_timeout_ledger, new_timeout);
-    }
-
-    #[test]
-    fn test_freelancer_requests_extension() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let (client, contract_client, freelancer, token_id) = setup(&env);
-        let job_id = String::from_str(&env, "ext-2");
-        let timeout_ledgers = 10u32;
-
-        client.create_escrow(
-            &job_id,
-            &contract_client,
-            &CreateEscrowParams {
-                freelancer: freelancer.clone(),
-                token: token_id.clone(),
-                amount: 500,
-                milestones: None,
-                timeout_ledgers: Some(timeout_ledgers),
-                referrer: None,
-            },
-        );
-
-        let escrow = client.get_escrow(&job_id);
-        let new_timeout = escrow.timeout_ledger + 20;
-
-        client.request_extension(&job_id, &freelancer, &new_timeout);
-
-        let req = client.get_extension_request(&job_id).unwrap();
-        assert_eq!(req.requested_by, freelancer);
-        assert_eq!(req.new_timeout_ledger, new_timeout);
-    }
-
-    #[test]
-    fn test_approve_extension_updates_timeout() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let (client, contract_client, freelancer, token_id) = setup(&env);
-        let job_id = String::from_str(&env, "ext-3");
-        let timeout_ledgers = 10u32;
-
-        client.create_escrow(
-            &job_id,
-            &contract_client,
-            &CreateEscrowParams {
-                freelancer: freelancer.clone(),
-                token: token_id.clone(),
-                amount: 500,
-                milestones: None,
-                timeout_ledgers: Some(timeout_ledgers),
-                referrer: None,
-            },
-        );
-
-        let escrow = client.get_escrow(&job_id);
-        let new_timeout = escrow.timeout_ledger + 50;
-
-        client.request_extension(&job_id, &contract_client, &new_timeout);
-        client.approve_extension(&job_id, &freelancer);
-
-        let updated = client.get_escrow(&job_id);
-        assert_eq!(updated.timeout_ledger, new_timeout);
-        assert!(client.get_extension_request(&job_id).is_none());
-    }
-
-    #[test]
-    fn test_freelancer_requests_client_approves() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let (client, contract_client, freelancer, token_id) = setup(&env);
-        let job_id = String::from_str(&env, "ext-4");
-        let timeout_ledgers = 10u32;
-
-        client.create_escrow(
-            &job_id,
-            &contract_client,
-            &CreateEscrowParams {
-                freelancer: freelancer.clone(),
-                token: token_id.clone(),
-                amount: 500,
-                milestones: None,
-                timeout_ledgers: Some(timeout_ledgers),
-                referrer: None,
-            },
-        );
-
-        let escrow = client.get_escrow(&job_id);
-        let new_timeout = escrow.timeout_ledger + 30;
-
-        client.request_extension(&job_id, &freelancer, &new_timeout);
-        client.approve_extension(&job_id, &contract_client);
-
-        let updated = client.get_escrow(&job_id);
-        assert_eq!(updated.timeout_ledger, new_timeout);
-    }
-
-    #[test]
-    fn test_extension_after_start_work() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let (client, contract_client, freelancer, token_id) = setup(&env);
-        let job_id = String::from_str(&env, "ext-5");
-        let timeout_ledgers = 10u32;
-
-        client.create_escrow(
-            &job_id,
-            &contract_client,
-            &CreateEscrowParams {
-                freelancer: freelancer.clone(),
-                token: token_id.clone(),
-                amount: 500,
-                milestones: None,
-                timeout_ledgers: Some(timeout_ledgers),
-                referrer: None,
-            },
-        );
-        client.start_work(&job_id, &freelancer);
-
-        let escrow = client.get_escrow(&job_id);
-        assert_eq!(escrow.status, EscrowStatus::InProgress);
-
-        let new_timeout = escrow.timeout_ledger + 20;
-        client.request_extension(&job_id, &freelancer, &new_timeout);
-        client.approve_extension(&job_id, &contract_client);
-
-        let updated = client.get_escrow(&job_id);
-        assert_eq!(updated.timeout_ledger, new_timeout);
-    }
-
-    #[test]
-    #[should_panic(expected = "An extension request is already pending")]
-    fn test_double_request_panics() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let (client, contract_client, freelancer, token_id) = setup(&env);
-        let job_id = String::from_str(&env, "ext-6");
-        let timeout_ledgers = 10u32;
-
-        client.create_escrow(
-            &job_id,
-            &contract_client,
-            &CreateEscrowParams {
-                freelancer: freelancer.clone(),
-                token: token_id.clone(),
-                amount: 500,
-                milestones: None,
-                timeout_ledgers: Some(timeout_ledgers),
-                referrer: None,
-            },
-        );
-
-        let escrow = client.get_escrow(&job_id);
-        let new_timeout = escrow.timeout_ledger + 50;
-        client.request_extension(&job_id, &contract_client, &new_timeout);
-        client.request_extension(&job_id, &contract_client, &(new_timeout + 10));
-    }
-
-    #[test]
-    #[should_panic(expected = "Cannot approve your own extension request")]
-    fn test_cannot_approve_own_request() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let (client, contract_client, freelancer, token_id) = setup(&env);
-        let job_id = String::from_str(&env, "ext-7");
-        let timeout_ledgers = 10u32;
-
-        client.create_escrow(
-            &job_id,
-            &contract_client,
-            &CreateEscrowParams {
-                freelancer: freelancer.clone(),
-                token: token_id.clone(),
-                amount: 500,
-                milestones: None,
-                timeout_ledgers: Some(timeout_ledgers),
-                referrer: None,
-            },
-        );
-
-        let escrow = client.get_escrow(&job_id);
-        let new_timeout = escrow.timeout_ledger + 50;
-        client.request_extension(&job_id, &contract_client, &new_timeout);
-        client.approve_extension(&job_id, &contract_client);
-    }
-
-    #[test]
-    #[should_panic(expected = "New timeout must be later than current timeout")]
-    fn test_new_timeout_must_be_later() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let (client, contract_client, freelancer, token_id) = setup(&env);
-        let job_id = String::from_str(&env, "ext-8");
-        let timeout_ledgers = 10u32;
-
-        client.create_escrow(
-            &job_id,
-            &contract_client,
-            &CreateEscrowParams {
-                freelancer: freelancer.clone(),
-                token: token_id.clone(),
-                amount: 500,
-                milestones: None,
-                timeout_ledgers: Some(timeout_ledgers),
-                referrer: None,
-            },
-        );
-
-        let escrow = client.get_escrow(&job_id);
-        client.request_extension(&job_id, &contract_client, &escrow.timeout_ledger);
-    }
-
-    #[test]
-    #[should_panic(expected = "Cannot extend timeout in current status")]
-    fn test_extension_on_released_panics() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let (client, contract_client, freelancer, token_id) = setup(&env);
-        let job_id = String::from_str(&env, "ext-9");
-        let timeout_ledgers = 10u32;
-
-        client.create_escrow(
-            &job_id,
-            &contract_client,
-            &CreateEscrowParams {
-                freelancer: freelancer.clone(),
-                token: token_id.clone(),
-                amount: 500,
-                milestones: None,
-                timeout_ledgers: Some(timeout_ledgers),
-                referrer: None,
-            },
-        );
-        client.release_escrow(&job_id, &contract_client);
-
-        let escrow = client.get_escrow(&job_id);
-        client.request_extension(&job_id, &contract_client, &(escrow.timeout_ledger + 10));
-    }
-
-    #[test]
-    fn test_extension_request_getter_returns_none_when_empty() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let (client, contract_client, freelancer, token_id) = setup(&env);
-        let job_id = String::from_str(&env, "ext-10");
-        let timeout_ledgers = 10u32;
-
-        client.create_escrow(
-            &job_id,
-            &contract_client,
-            &CreateEscrowParams {
-                freelancer: freelancer.clone(),
-                token: token_id.clone(),
-                amount: 500,
-                milestones: None,
-                timeout_ledgers: Some(timeout_ledgers),
-                referrer: None,
-            },
-        );
-
-        assert!(client.get_extension_request(&job_id).is_none());
-    }
-
-    #[test]
-    #[should_panic(expected = "No pending extension request")]
-    fn test_approve_without_request_panics() {
-        let env = Env::default();
-        env.mock_all_auths();
-        let (client, contract_client, freelancer, token_id) = setup(&env);
-        let job_id = String::from_str(&env, "ext-11");
-        let timeout_ledgers = 10u32;
-
-        client.create_escrow(
-            &job_id,
-            &contract_client,
-            &CreateEscrowParams {
-                freelancer: freelancer.clone(),
-                token: token_id.clone(),
-                amount: 500,
-                milestones: None,
-                timeout_ledgers: Some(timeout_ledgers),
-                referrer: None,
-            },
-        );
-
-        client.approve_extension(&job_id, &freelancer);
-    }
-}
-#[cfg(test)]
-mod usdc_tests {
-    use super::*;
-    use crate::token;
-    use soroban_sdk::{testutils::Address as _, Address, Env, String};
-
-    fn setup(env: &Env) -> (MarketPayContractClient, Address, Address, Address) {
-        env.mock_all_auths();
-        let id = env.register(MarketPayContract, ());
-        let contract = MarketPayContractClient::new(env, &id);
-        let admin = Address::generate(env);
-        let treasury = Address::generate(env);
-        contract.initialize(&admin, &treasury);
-
-        let client = Address::generate(env);
-        let freelancer = Address::generate(env);
-
-        // Mocking a generic token contract which will simulate USDC
-        let usdc_contract = env.register_stellar_asset_contract_v2(admin.clone());
-        let usdc_id = usdc_contract.address();
-        let usdc_admin = token::StellarAssetClient::new(env, &usdc_id);
-        usdc_admin.mint(&client, &100_000_000); // Give the client some USDC
-
-        (contract, client, freelancer, usdc_id)
-    }
-
-    #[test]
-    fn test_usdc_escrow_flow() {
-        let env = Env::default();
-        let (contract, client, freelancer, usdc_id) = setup(&env);
-        let job_id = String::from_str(&env, "usdc-job-1");
-
-        let token_client = token::TokenClient::new(&env, &usdc_id);
-
-        // Initial balances
-        assert_eq!(token_client.balance(&client), 100_000_000);
-        assert_eq!(token_client.balance(&contract.address), 0);
-        assert_eq!(token_client.balance(&freelancer), 0);
-
-        // 1. Create escrow with USDC
-        contract.create_escrow(
-            &job_id,
-            &client,
-            &CreateEscrowParams {
-                freelancer: freelancer.clone(),
-                token: usdc_id.clone(),
-                amount: 5_000,
-                milestones: None,
-                timeout_ledgers: None,
-                referrer: None,
-            },
-        );
-
-        // Verify contract holds USDC
-        assert_eq!(token_client.balance(&client), 99_995_000);
-        assert_eq!(token_client.balance(&contract.address), 5_000);
-
-        // 2. Freelancer starts work
-        contract.start_work(&job_id, &freelancer);
-
-        // 3. Client releases escrow
-        contract.release_escrow(&job_id, &client);
-
-        // Verify freelancer received the funds (assuming 0% fee here or similar, but the exact amount doesn't matter as long as it's not 0 and the contract is 0)
-        let final_freelancer_balance = token_client.balance(&freelancer);
-        assert!(final_freelancer_balance > 0);
-        assert_eq!(token_client.balance(&contract.address), 0);
-    }
-}
+mod tests;
