+++ contracts/marketpay-contract/src/tests/mod.rs
@@ -1,0 +1,10 @@
+mod basic_tests;
+mod deliverable_hash_tests;
+mod deliverable_oracle_tests;
+mod extension_tests;
+mod milestone_pct_tests;
+mod regression_tests;
+mod sealed_bid_tests;
+mod timeout_tests;
+mod upgrade_tests;
+mod usdc_tests;
